% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident}
\alias{accident}
\title{도로교통공단 사망교통사고 정보}
\format{
A data frame with 23 variables.
\describe{
  \item{발생년}{발생년}
  \item{발생년월일시}{발생년월일시}
  \item{주야}{주야}
  \item{요일}{요일}
  \item{사망자수}{사망자수}
  \item{부상자수}{부상자수}
  \item{중상자수}{중상자수}
  \item{경상자수}{경상자수}
  \item{부상신고자수}{부상신고자수}
  \item{발생지시도}{발생지시도}
  \item{발생지시군구}{발생지시군구}
  \item{사고유형_대분류}{사고유형_대분류}
  \item{사고유형_중분류}{사고유형_중분류}
  \item{사고유형}{사고유형}
  \item{가해자법규위반}{가해자법규위반}
  \item{도로형태_대분류}{도로형태_대분류}
  \item{도로형태}{도로형태}
  \item{가해자_당사자종별}{가해자_당사자종별}
  \item{피해자_당사자종별}{피해자_당사자종별}
  \item{발생위치X_UTMK}{발생위치X_UTMK}
  \item{발생위치Y_UTMK}{발생위치Y_UTMK}
  \item{경도}{경도}
  \item{위도}{위도}
 }
}
\usage{
accident
}
\description{
도로교통공단에서 제공하는 사망 교통사고에 대한 개별정보(발생일시, 사고유형, 위치좌표 등)
부상자수 = 중상자수 + 경상자수 + 부상신고자수
https://www.data.go.kr/data/15070340/fileData.do
}
\examples{
 library(leaflet)
 accident |> 
   dplyr::filter(발생지시도 == "서울") |> 
   dplyr::filter(사망자수 > 0) |> 
   dplyr::filter(피해자_당사자종별 == "보행자") |> 
   leaflet() |> 
   addTiles() |> 
   addMarkers(~경도, ~위도, popup = ~as.character(사고유형), label = ~as.character(사고유형))
}
\keyword{datasets}
