% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accom}
\alias{accom}
\title{행정안전부 관광숙박업 정보}
\format{
A data frame with 16 variables:
\describe{
  \item{사업장명}{해당 관광숙박업 사업장명}
  \item{영업상태명}{해당 관광숙박업 영업상태명 : 취소/말소/만료/정지/중지, 영업/정상,폐업,휴업}
  \item{폐업일자}{해당 관광숙박업 폐업일자}
  \item{전화번호}{해당 관광숙박업 전화번호}
  \item{소재지면적}{해당 관광숙박업 소재지면적}
  \item{소재지우편번호}{해당 관광숙박업 소재지우편번호}
  \item{소재지전체주소}{해당 관광숙박업 소재지전체주소}
  \item{도로명전체주소}{해당 관광숙박업 도로명전체주소}
  \item{도로명우편번호}{해당 관광숙박업 도로명우편번호}
  \item{경도}{해당 관광숙박업 경도}
  \item{위도}{해당 관광숙박업 위도}
  \item{총층수}{해당 관광숙박업 총층수}
  \item{지상층수}{해당 관광숙박업 지상층수}
  \item{지하층수}{해당 관광숙박업 지하층수}
  \item{객실수}{해당 관광숙박업 객실수}
  \item{영문상호명}{해당 관광숙박업 영문상호명}
}
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
accom
}
\description{
관광호텔과 같은 숙박시설 및 음식, 운동, 오락 등 적합한 시설 등을 함께 갖추어 제공하는 업소정보, 주소, 번호, 경도, 위도, 객실수 등을 제공
}
\examples{
library(leaflet)
leaflet(accom) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=8) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~`경도`, lat=~`위도`,
             label =~`사업장명`,
             popup =~paste(`사업장명`,"<BR>", "주소 : ",`도로명전체주소`, sep="","<BR>","전화번호 : ",`전화번호`))
}
\keyword{datasets}
