% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airport}
\alias{airport}
\title{국토교통부 전세계 공항정보}
\format{
A list with 8 data frame.
\describe{
   \item{영문공항명}{영문공항명}
   \item{한글공항명}{한글공항명}
   \item{공항코드1.IATA.}{공항코드1 IATA}
   \item{공항코드2.ICAO.}{공항코드2 ICAO}
   \item{지역}{유럽, 중남미, 아프리카, 중동, 북미, 아시아, 대양주, 남미}
   \item{영문국가명}{영문국가명}
   \item{한글국가명}{한글국가명}
   \item{영문도시명}{영문도시명}
 }
}
\usage{
airport
}
\description{
국토교통부에서 제공하는 전세계 공항의 영문 공항명, 한글 공항명, IATA 코드, ICAO 코드, 국가명, 지역명에 대한 정보.
Last Update: 2022-06-10;
https://www.data.go.kr/data/3051587/fileData.do
}
\keyword{datasets}
