% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bike}
\alias{bike}
\title{서울특별시 자전거편의시설}
\format{
A data frame with 6 variables:
\describe{
  \item{편의시설명}{자전거 편의 시설 명(ID)}
  \item{편의시설분류}{자전거 편의시설 분류(또는 위치)}
  \item{경도}{해당 자전거 편의시설 경도}
  \item{위도}{해당 자전거 편의시설 위도}
  \item{주소}{해당 자전거 편의시설 주소}
  \item{우편번호}{해당 자전거 편의시설 우편번호}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-21291/S/1/datasetView.do}
}
\usage{
bike
}
\description{
서울시 자전거 편의시설현황입니다. 편의시설분류, 위도, 경도, 주소, 우편번호의 정보를 제공합니다.
}
\examples{
library(leaflet)
leaflet(bike) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=11) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~경도, lat=~위도,
             label=~편의시설명,
             popup=~paste(편의시설명, "<BR>", "주소 : ",주소))
}
\keyword{datasets}
