% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget2023}
\alias{budget2023}
\title{기획재정부 연도별 세출 및 지출 예산현황}
\format{
A data frame with 13 variables.
\describe{
   \item{회계연도}{회계상의 정리기술면에서 편의에 따라 일정한 기간을 정하고 그 수지상황을 명확히 하는 예산의 기간적 단위. 우리나라의 경우에 현행 국가재정법법 제2조에 따라 매년 1월1일에 시작하여 12월 31일에 종료.}
   \item{소관명}{국가의 업무 담당하여 관리하는 부, 처, 청, 위원회 등 정부조직법상 중앙행정기관과 헌법기관의 이름}
   \item{회계명}{정보를 제공하기 위한 일련의 과정 또는 체계. 일반회계는 조세수입 등을 주요 세입으로 하여 국가의 일반적인 세출에 충당하기 위하여 설치하며 특별회계는 국가에서 특정한 사업을 운영하고자 할 때, 특정한 자금을 보유하여 운용하고자 할 때, 특정한 세입으로 특정한 세출에 충당함으로써 일반회계와 구분하여 회계처리할 필요가 있을 때에 법률로써 설치}
   \item{계정명}{회계에 있어서 재산의 모든 변동을 종합적, 유기적으로 파악하여 기록하고 계산하기 위한 특수형식.국가재정에서는 일반적인 의미의 계정(account, 계정과목)이외에 세입세출 예산을 구분하기위한 계정이 존재(국가재정법 제21조)}
   \item{분야명}{프로그램 예산제도에 따른 세출예산 분류 체계(분야-부문-프로그램-단위사업-세부사업)의 최상위 단위}
   \item{부문명}{프로그램 예산제도에 따른 세출예산 분류 체계(분야-부문-프로그램-단위사업-세부사업)의 한 단위}
   \item{프로그램명}{프로그램 예산제도에 따른 세출예산 분류 체계(분야-부문-프로그램-단위사업-세부사업)의 한 단위. 국가의 최소 정책단위로서 동일한 정책목표를 달성하기 위한 한 개 이상의 단위사업으로 구성}
   \item{단위사업명}{프로그램 예산제도에 따른 세출예산 분류 체계(분야-부문-프로그램-단위사업-세부사업)의 한 단위로 프로그램(정책) 달성을 위한 세부사업군(群)으로 구성 (예: 수산물 가격 안정)}
   \item{세부사업명}{프로그램 예산제도에 따른 세출예산 분류 체계(분야-부문-프로그램-단위사업-세부사업)의 최하위 단위로 행정과목에 해당}
   \item{경비구분}{인건비, 기본경비 등 경비 구분}
   \item{지출구분}{일반지출, 보전 지출 등 지출 구분}
   \item{정부안금액_천원}{각 중앙관서의 장이 제출한 예산 요구서에 대한 검토 및 조정 작업(예산안 편성)을 거쳐 예산안 시안이 마련되면 기획재정부는 대통령 중간보고, 당정협의, 예산자문회의 등을 거쳐 정부의 최종 예산안을 마련}
   \item{국회확정금액_천원}{국회에서 확정된 최종 금액}
 }
}
\source{
\url{https://www.data.go.kr/data/15071311/fileData.do}
}
\usage{
budget2023
}
\description{
2023년 정부 연도별 세출/지출 예산편성 현황(총액), 기금운용계획안. 이용허락범위 제한 없음.
}
\examples{
DT::datatable(head(budget2023))
}
\keyword{datasets}
