% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{busyMetro}
\alias{busyMetro}
\title{서울교통공사 지하철혼잡도정보}
\format{
A data frame with 7 variables.
\describe{
   \item{요일구분}{평일, 토요일, 공휴일 등 요일 구분}
   \item{호선}{지하철 호선 번호}
   \item{역번호}{외부역 코드}
   \item{출발역}{출발역 이름}
   \item{상하구분}{운행 방향}
   \item{시간}{기준 시간으로 30분 단위로 집계}
   \item{혼잡도}{30분 단위 평균 혼잡도로 30분간 지나는 열차들의 평균 혼잡도}
 }
}
\source{
\url{https://www.data.go.kr/data/15071311/fileData.do}
}
\usage{
busyMetro
}
\description{
서울교통공사 1-8호선 30분 단위 평균 혼잡도로 30분간 지나는 열차들의 평균 혼잡도(정원대비 승차인원으로, 승차인과 좌석수가 일치할 경우를 혼잡도 34%로 산정) 제공(단위: %). 요일구분(평일, 토요일, 일요일), 호선, 역번호, 역명, 상하선구분, 30분단위 별 혼잡도 데이터로 구성되어 있음. 1년 단위 업데이트. 이용허락범위 제한없음.
}
\examples{
DT::datatable(head(busyMetro))

busyMetro \%>\% 
   filter(요일구분 == "평일") \%>\% 
   filter(호선 == 2) \%>\% 
   mutate(출발역 = paste(호선, "호선", 출발역)) \%>\% 
   ggplot(aes(시간, 혼잡도, color = 상하구분)) +
   geom_line(aes(group = 상하구분)) +
   scale_color_manual(values = c("#E966A0", "#213363", "#8EAC50", "#F2BE22")) +
   facet_wrap(. ~출발역, ncol = 10, strip.position = "bottom") +
   ggthemes::theme_tufte(base_family = "NanumGothic") +
   theme(
       axis.text = element_blank(),
       axis.ticks = element_blank(),
       axis.title = element_blank(),
       legend.position = "bottom"
       ) +
   labs(title = "서울교통공사 2호선 시간대별 지하철 혼잡도 분포", subtitle = "평일 05:30 ~ 00:30")
}
\keyword{datasets}
