% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carInspection}
\alias{carInspection}
\title{한국교통안전공단 자동차검사소 정보}
\format{
A data frame with 22 variables.
\describe{
   \item{자동차검사소명}{자동차검사소명}
   \item{자동차검사소유형}{자동차검사소유형}
   \item{소재지도로명주소}{소재지도로명주소}
   \item{소재지지번주소}{소재지지번주소}
   \item{위도}{위도}
   \item{경도}{경도}
   \item{검사소전화번호}{검사소전화번호}
   \item{운영시간}{운영시간}
   \item{검사진로수}{검사진로수}
   \item{부지면적}{부지면적}
   \item{건물면적}{건물면적}
   \item{검사기술인력수}{검사기술인력수}
   \item{신규검사여부}{Y, N}
   \item{정기검사여부}{Y, N}
   \item{튜닝검사여부}{Y, N}
   \item{임시검사여부}{Y, N}
   \item{수리검사여부}{Y, N}
   \item{배출가스정밀검사여부}{Y, N}
   \item{택시미터검정여부}{Y, N}
   \item{관리기관전화번호}{관리기관전화번호}
   \item{관리기관명}{관리기관명}
   \item{데이터기준일자}{데이터기준일자}
 }
}
\usage{
carInspection
}
\description{
한국교통안전공단에서 운영하고 있는 자동차검사소 정보. 해당 지역별 리스트와 위도, 경도 등 각종 정보를 제공.
Last Update: 2021-09-27;
https://www.data.go.kr/data/3041572/fileData.do
}
\keyword{datasets}
