% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childAbuse}
\alias{childAbuse}
\title{보건복지부 아동학대 신고정보}
\format{
A data frame with 4 variables.
\describe{
   \item{기간}{아동학대 신고 기간
   년도별, 2013년~2018년}
   \item{구분}{아동학대 신고 지역 구분}
   \item{신고자}{아동학대 신고자 구분}
   \item{신고건수}{아동학대 신고건수}
 }
}
\source{
\url{https://www.data.go.kr/data/15065291/fileData.do}{공공데이터포털:사회복지 - 보육·가족및여성}
}
\usage{
childAbuse
}
\description{
보건복지부 아동학대 신고정보에 대한 해당연도 말 시점의 시도별, 의심사례, 일반상담, 동일신고, 
해외발생, 교직원, 의료인 등 신고현황통계.
Last Update: 2022-06-30;
}
\examples{
DT::datatable(head(childAbuse,50))
}
\keyword{datasets}
