% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crimePlace}
\alias{crimePlace}
\title{경찰청 범죄 발생 장소별 통계}
\format{

}
\source{
\url{https://www.data.go.kr/data/3074463/fileData.do}{공공데이터포털:공공질서및안전 - 경찰}
}
\usage{
crimePlace
}
\description{
경찰청에서 제공하는 범죄유형별(강력범죄, 폭력범죄, 지능범죄 등)로 범죄장소(아파트, 단독주택, 고속도로, 노상 ,백화점, 슈퍼마켓 등)별 통계. 이용허락범위 제한 없음.
Last Update: 2022-10-08;
}
\examples{
DT::datatable(head(crimePlace,50))
}
\keyword{datasets}
