% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economyPeople}
\alias{economyPeople}
\title{통계청 경제활동인구조사}
\format{
A data frame with 7 variables:
\describe{
  \item{연령계층별}{해당 시점의 경제활동인구 조사 연령계층 : 15-19세, 20-29세, 30-39세, 40-49세, 50-59세, 60세 이상}
  \item{시점}{경제활동인구를 실시한 시점}
  \item{인구 (천명)}{해당 연령계층의 총 인구}
  \item{경제활동인구 (천명)}{만 15세 이상 인구 중 취업자와 실업자를 말한다}
  \item{취업자 (천명)}{① 조사대상 주간 중 수입을 목적으로 1시간 이상 일한 자, ② 자기에게 직접적으로는 이득이나 수입이 오지 않더라도 자기가구에서 경영하는 농장이나 사업체의 수입을 높이는 데 도운 가족종사자로서 주당 18시간이상 일한 자(무급가족종사자), ③ 직장 또는 사업체를 가지고 있으나 조사대상 주간 중 일시적인 병, 일기불순, 휴가 또는 연가, 노동쟁의 등의 이유로 일하지 못한 일시휴직자}
  \item{실업자 (천명)}{조사대상주간에 수입 있는 일을 하지 않았고, 지난 4주간 일자리를 찾아 적극적으로 구직활동을 하였던 사람으로서 일자리가 주어지면 즉시 취업이 가능한 사람}
  \item{비경제활동인구 (천명)}{조사대상 주간 중 취업자도 실업자도 아닌 만 15세 이상인 자, 즉 집안에서 가사와 육아를 전담하는 가정주부, 학교에 다니는 학생, 일을 할 수 없는 연로자와 심신장애자, 자발적으로 자선사업이나 종교단체에 관여하는 자 등을 말한다}
}
}
\source{
\url{https://kosis.kr/statHtml/statHtml.do?orgId=101&tblId=DT_1DA7002S&vw_cd=MT_ZTITLE&list_id=B11&seqNo=&lang_mode=ko&language=kor&obj_var_id=&itm_id=&conn_path=MT_ZTITLE}
}
\usage{
economyPeople
}
\description{
국민의 경제활동(취업, 실업, 노동력 등) 특성을 조사함으로써 거시경제 분석과 인력자원의 개발정책 수립에 필요한 기초 자료를 제공
}
\examples{
library(dplyr)
library(ggplot2)
library(gganimate)

ani_economyPeople <- economyPeople \%>\% 
  mutate(
    `경제활동인구비율` = `경제활동인구 (천명)` / `인구 (천명)` * 100
  ) \%>\% 
  ungroup() \%>\% 
  arrange(시점, 연령계층별) \%>\% 
  ungroup()

plot_ecoPeople <- ggplot(
  ani_economyPeople,
  aes(x = `연령계층별`, y = `경제활동인구비율`, fill = `연령계층별`)) +
  theme_minimal(base_family = "AppleSDGothicNeo-SemiBold") +
  geom_bar(stat='identity') + 
  theme(
    axis.text.x = element_text(size = 15, color = "grey3", face = "bold"),
    axis.title = element_text(size = 17, color = "grey21", face = "bold"),
    legend.position = "none",
    plot.title = element_text(hjust = 0.5, size = 22, color = "royalblue4", face = "bold"),
    axis.line = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.grid.major.x = element_line(size = 0.1, color = "grey"),
    plot.background = element_blank()
  ) + 
  labs(
    title = "{closest_state} 연령별 경제활동인구비율",
    subtitle = "1999-2023년",
    caption = "KOSTAT; 경제활동인구조사(통계청)",
    x = "",
    y = "경제활동인구비율"
  ) +
  transition_states(
    시점
  ) 

animate(plot_ecoPeople, duration = 30)
}
\keyword{datasets}
