% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election2020}
\alias{election2020}
\title{중앙선거관리위원회 국회의원선거 개표결과 정보}
\format{
A data frame with 8 variables.
\describe{
   \item{구분}{비례대표, 국회의원 및 보궐선거}
   \item{지역구명}{시군구 지역구 이름}
   \item{읍면동명}{읍면동 이름}
   \item{투표구명}{투표구 이름}
   \item{선거인수}{선거권이 있는 사람으로서 선거인명부 또는 재외선거인명부에 올라 있는 사람의 수}
   \item{투표수}{선거인 중 투표에 참여한 사람의 수}
   \item{정당}{투표 정당 및 후보자 이름}
   \item{득표}{득표 수}
 }
}
\source{
\url{https://www.data.go.kr/data/15025527/fileData.do}
}
\usage{
election2020
}
\description{
2020년도 전국 시도, 구시군, 읍면동, 투표구별 국회의원선거의 개표정보. 이용허락범위 제한 없음.
}
\examples{
DT::datatable(head(election2020))
}
\keyword{datasets}
