% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elevator}
\alias{elevator}
\title{한국승강기안전공단 국내 승강기 보유 현황}
\format{
A list with 8 data frame.
}
\usage{
elevator
}
\description{
한국승강기안전공단에서 제공하는 국내 승강기 보유 현황. 이용허락범위 제한 없음.
Last Update: 2022-02-22;
https://www.data.go.kr/data/15017366/fileData.do
}
\keyword{datasets}
