% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farmGIS}
\alias{farmGIS}
\title{행전안전부 가축사육업 로컬데이터}
\format{
A data frame with 34 variables.
\describe{
   \item{}{}
 }
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
farmGIS
}
\description{
가축을 사육하여 판매하거나 젖·알·꿀을 생산하는 업소정보	. 이용허락범위 제한 없음.
}
\examples{
DT::datatable(head(farmGIS))
library(sf)
library(tidyverse)
library(leaflet)
library(inlmisc)

farm <- farmGIS \%>\% 
  separate(col = "도로명전체주소", into = c("시도", "시군구"), sep = " ") \%>\% 
  filter(주사육업종 == "한우") \%>\%
  filter(영업상태명 == "영업/정상") \%>\% 
  mutate(
    x = as.numeric(좌표정보_x),
    y = as.numeric(좌표정보_y),
    idx = row_number()
  ) \%>\% 
  select(사업장명, x, y, idx) \%>\% 
  filter(!is.na(y)) \%>\% 
  st_as_sf(coords = c("x", "y"), crs = 5174) \%>\% 
  st_transform(4326)

leaflet() \%>\% 
  addTiles() \%>\% 
  addMarkers(data = farm, label = ~as.character(사업장명), clusterOptions = markerClusterOptions(), group="marker") \%>\% 
  inlmisc::AddSearchButton(group = "marker", zoom = 15, textPlaceholder = "Search here")
}
\keyword{datasets}
