% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{소방청 화재통계}
\format{
A list with 18 data frame.
}
\usage{
fire
}
\description{
소방청에서 제공하는 장소별 원인별 화재통계.
Last Update: 2020-03-19;
https://www.data.go.kr/data/3038724/fileData.do
}
\keyword{datasets}
