% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foodNutrients}
\alias{foodNutrients}
\title{식품의약품안전처 식품영양성분 데이터베이스}
\format{
A list with 146 data frame.
\describe{
   \item{no}{순번}
   \item{sample_id}{sample_id}
   \item{식품코드}{해당식품의 식품코드}
   \item{db_군}{농축산물}
   \item{상용제품}{해당식품의 상용제품
   품목대표}
   \item{식품명}{해당식품의 식품명}
   \item{연도}{해당식품의 조사 연도
   2019, 2020, 2021}
   \item{지역_제조사}{해당식품의 제조사 지역
   전국(대표)}
   \item{채취시기}{해당식품의 채취시기
   평균}
   \item{식품대분류}{해당식품의 식품대분류
   농축산물}
   \item{식품상세분류}{해당식품의 식품상세분류
   곡류 및 그 제품, 감자 및 전분류, 당류, 두류, 견과류 및 종실류, 채소류, 버섯류, 과실류, 육류, 난류, 어패류 및 기타 수산물, 해조류, 우유 및 유제품류, 유지류, 차류, 음료류, 주류, 조미료류, 조리가공품류, 기타}
   \item{x1회제공량}{해당식품의 1회 제공량
   100}
   \item{내용량_단위}{해당식품 내용량의 단위 : g}
   \item{에너지}{해당식품의 에너지(1회제공량당)}
   \item{수분_g}{해당식품의 수분(1회제공량당)}
   \item{단백질_g}{해당식품의 단백질_g(1회제공량당)}
   \item{지방_g}{해당식품의 지방_g(1회제공량당)}
   \item{탄수화물_g}{해당식품의 탄수화물_g(1회제공량당)}
   \item{총당류_g}{해당식품의 총당류_g(1회제공량당)}
   \item{자당_g}{해당식품의 자당_g(1회제공량당)}
   \item{포도당_g}{해당식품의 포도당_g(1회제공량당)}
   \item{과당_g}{해당식품의 과당_g(1회제공량당)}
   \item{유당_g}{해당식품의 유당_g(1회제공량당)}
   \item{맥아당_g}{해당식품의 맥아당_g(1회제공량당)}
   \item{갈락토오스_g}{해당식품의 갈락토오스_g(1회제공량당)}
   \item{총_식이섬유_g}{해당식품의 총 식이섬유_g(1회제공량당)}
   \item{수용성_식이섬유_g}{해당식품의 수용성 식이섬유_g(1회제공량당)}
   \item{불용성_식이섬유_g}{해당식품의 불용성 식이섬유_g(1회제공량당)}
   \item{칼슘}{해당식품의 칼슘(1회제공량당)}
   \item{철}{해당식품의 철(1회제공량당)}
   \item{마그네슘}{해당식품의 마그네슘(1회제공량당)}
   \item{인}{해당식품의 인(1회제공량당)}
   \item{칼륨}{해당식품의 칼륨(1회제공량당)}
   \item{나트륨}{해당식품의 나트륨(1회제공량당)}
   \item{아연}{해당식품의 아연(1회제공량당)}
   \item{구리}{해당식품의 구리(1회제공량당)}
   \item{망간}{해당식품의 망간(1회제공량당)}
   \item{셀레늄}{해당식품의 셀레늄(1회제공량당)}
   \item{몰리브덴}{해당식품의 몰리브덴(1회제공량당)}
   \item{요오드}{해당식품의 요오드(1회제공량당)}
   \item{레티놀}{해당식품의 레티놀(1회제공량당)}
   \item{베타카로틴}{해당식품의 베타카로틴(1회제공량당)}
   \item{비타민_d_d2_d3}{해당식품의 비타민 d_d2_d3(1회제공량당)}
   \item{비타민_d2}{해당식품의 비타민 d2(1회제공량당)}
   \item{비타민_d3}{해당식품의 비타민 d3(1회제공량당)}
   \item{비타민_e}{해당식품의 비타민 e(1회제공량당)}
   \item{알파_토코페롤}{해당식품의 알파 토코페롤(1회제공량당)}
   \item{베타_토코페롤}{해당식품의 베타 토코페롤(1회제공량당)}
   \item{감마_토코페롤}{해당식품의 감마 토코페롤(1회제공량당)}
   \item{델타_토코페롤}{해당식품의 델타 토코페롤(1회제공량당)}
   \item{알파_토코트리에놀}{해당식품의 알파 토코트리에놀(1회제공량당)}
   \item{베타_토코트리에놀}{해당식품의 베타 토코트리에놀(1회제공량당)}
   \item{감마_토코트리에놀}{해당식품의 감마 토코트리에놀(1회제공량당)}
   \item{델타_토코트리에놀}{해당식품의 델타 토코트리에놀(1회제공량당)}
   \item{비타민_k}{해당식품의 비타민 k(1회제공량당)}
   \item{비타민_k1}{해당식품의 비타민 k1(1회제공량당)}
   \item{비타민_k2}{해당식품의 비타민 k2(1회제공량당)}
   \item{비타민_b1}{해당식품의 비타민 b1(1회제공량당)}
   \item{비타민_b2}{해당식품의 비타민 b2(1회제공량당)}
   \item{나이아신}{해당식품의 나이아신(1회제공량당)}
   \item{나이아신_ne}{해당식품의 나이아신_ne(1회제공량당)}
   \item{니코틴산}{해당식품의 니코틴산(1회제공량당)}
   \item{니코틴아마이드}{해당식품의 니코틴아마이드(1회제공량당)}
   \item{판토텐산}{해당식품의 판토텐산(1회제공량당)}
   \item{비타민_b6}{해당식품의 비타민 b6(1회제공량당)}
   \item{피리독신}{해당식품의 피리독신(1회제공량당)}
   \item{비오틴}{해당식품의 비오틴(1회제공량당)}
   \item{엽산_dfe}{해당식품의 엽산 dfe(1회제공량당)}
   \item{엽산_천연_엽산}{해당식품의 엽산 천연 엽산(1회제공량당)}
   \item{엽산_합성_엽산}{해당식품의 엽산 합성 엽산(1회제공량당)}
   \item{비타민_b12}{해당식품의 비타민 b12(1회제공량당)}
   \item{비타민_c}{해당식품의 비타민 c(1회제공량당)}
   \item{총_아미노산}{해당식품의 총 아미노산(1회제공량당)}
   \item{필수_아미노산}{해당식품의 필수 아미노산(1회제공량당)}
   \item{이소류신}{해당식품의 이소류신(1회제공량당)}
   \item{류신}{해당식품의 류신(1회제공량당)}
   \item{라이신}{해당식품의 라이신(1회제공량당)}
   \item{메티오닌}{해당식품의 메티오닌(1회제공량당)}
   \item{페닐알라닌}{해당식품의 페닐알라닌(1회제공량당)}
   \item{트레오닌}{해당식품의 트레오닌(1회제공량당)}
   \item{트립토판}{해당식품의 트립토판(1회제공량당)}
   \item{발린}{해당식품의 발린(1회제공량당)}
   \item{히스티딘}{해당식품의 히스티딘(1회제공량당)}
   \item{아르기닌}{해당식품의 아르기닌(1회제공량당)}
   \item{비필수_아미노산}{해당식품의 비필수 아미노산(1회제공량당)}
   \item{티로신}{해당식품의 티로신(1회제공량당)}
   \item{시스테인}{해당식품의 시스테인(1회제공량당)}
   \item{알라닌}{해당식품의 알라닌(1회제공량당)}
   \item{아스파르트산}{해당식품의 아스파르트산(1회제공량당)}
   \item{글루탐산}{해당식품의 글루탐산(1회제공량당)}
   \item{글리신}{해당식품의 글리신(1회제공량당)}
   \item{프롤린}{해당식품의 프롤린(1회제공량당)}
   \item{세린}{해당식품의 세린(1회제공량당)}
   \item{타우린}{해당식품의 타우린(1회제공량당)}
   \item{콜레스테롤}{해당식품의 콜레스테롤(1회제공량당)}
   \item{총_지방산_g}{해당식품의 총 지방산_g(1회제공량당)}
   \item{총_필수_지방산_g}{해당식품의 총 필수 지방산_g(1회제공량당)}
   \item{총_포화_지방산_g}{해당식품의 총 포화 지방산_g(1회제공량당)}
   \item{부티르산_4_0}{해당식품의 부티르산 4_0(1회제공량당)}
   \item{카프로산_6_0}{해당식품의 카프로산 6_0(1회제공량당)}
   \item{카프릴산_8_0}{해당식품의 카프릴산 8_0(1회제공량당)}
   \item{카프르산_10_0}{해당식품의 카프르산 10_0(1회제공량당)}
   \item{라우르산_12_0}{해당식품의 라우르산 12_0(1회제공량당)}
   \item{트라이데칸산_13_0}{해당식품의 트라이데칸산 13_0(1회제공량당)}
   \item{미리스트산_14_0}{해당식품의 미리스트산 14_0(1회제공량당)}
   \item{펜타데칸산_15_0}{해당식품의 펜타데칸산 15_0(1회제공량당)}
   \item{팔미트산_16_0}{해당식품의 팔미트산 16_0(1회제공량당)}
   \item{헵타데칸산_17_0}{해당식품의 헵타데칸산 17_0(1회제공량당)}
   \item{스테아르산_18_0}{해당식품의 스테아르산 18_0(1회제공량당)}
   \item{아라키드산_20_0}{해당식품의 아라키드산 20_0(1회제공량당)}
   \item{헨에이코산산_21_0}{해당식품의 헨에이코산산 21_0(1회제공량당)}
   \item{베헨산_22_0}{해당식품의 베헨산 22_0(1회제공량당)}
   \item{트리코산산_23_0}{해당식품의 트리코산산 23_0(1회제공량당)}
   \item{리그노세르산_24_0}{해당식품의 리그노세르산 24_0(1회제공량당)}
   \item{총_단일_불포화지방산_g}{해당식품의 총 단일 불포화지방산_g(1회제공량당)}
   \item{미리스톨레산_14_1}{해당식품의 미리스톨레산 14_1(1회제공량당)}
   \item{팔미톨레산_16_1}{해당식품의 팔미톨레산 16_1(1회제공량당)}
   \item{헵타데센산_17_1}{해당식품의 헵타데센산 17_1(1회제공량당)}
   \item{올레산_18_1_n_9}{해당식품의 올레산 18_1_n_9(1회제공량당)}
   \item{박센산_18_1_n_7}{해당식품의 박센산 18_1_n_7(1회제공량당)}
   \item{가돌레산_20_1}{해당식품의 가돌레산 20_1(1회제공량당)}
   \item{에루크산_22_1}{해당식품의 에루크산 22_1(1회제공량당)}
   \item{네르본산_24_1}{해당식품의 네르본산 24_1(1회제공량당)}
   \item{총_다중_불포화지방산_g}{해당식품의 총 다중 불포화지방산_g(1회제공량당)}
   \item{리놀레산_18_2_n_6_c}{해당식품의 리놀레산 18_2_n_6_c(1회제공량당)}
   \item{알파_리놀렌산_18_3_n_3}{해당식품의 알파 리놀렌산_18_3_n_3(1회제공량당)}
   \item{감마_리놀렌산_18_3_n_6}{해당식품의 감마 리놀렌산_18_3_n_6(1회제공량당)}
   \item{에이코사디에노산_20_2_n_6}{해당식품의 에이코사디에노산 20_2_n_6(1회제공량당)}
   \item{에이코사트리에노산_20_3_n_3}{해당식품의 에이코사트리에노산 20_3_n_3(1회제공량당)}
   \item{에이코사트리에노산_20_3_n_6}{해당식품의 에이코사트리에노산 20_3_n_6(1회제공량당)}
   \item{아라키돈산_20_4_n_6}{해당식품의 아라키돈산 20_4_n_6(1회제공량당)}
   \item{에이코사펜타에노산_20_5_n_3}{해당식품의 에이코사펜타에노산 20_5_n_3(1회제공량당)}
   \item{도코사디에노산_22_2}{해당식품의 도코사디에노산 22_2(1회제공량당)}
   \item{도코사펜타에노산_22_5_n_3}{해당식품의 도코사펜타에노산 22_5_n_3(1회제공량당)}
   \item{도코사헥사에노산_22_6_n_3}{해당식품의 도코사헥사에노산 22_6_n_3(1회제공량당)}
   \item{오메가_3_지방산_g}{해당식품의 오메가 3 지방산_g(1회제공량당)}
   \item{오메가_6_지방산_g}{해당식품의 오메가 6 지방산_g(1회제공량당)}
   \item{트랜스_지방산_g}{해당식품의 트랜스 지방산_g(1회제공량당)}
   \item{트랜스_올레산_18_1_n_9_t}{해당식품의 트랜스 올레산_18_1_n_9_t(1회제공량당)}
   \item{트랜스_리놀레산_18_2t}{해당식품의 트랜스 리놀레산_18_2t(1회제공량당)}
   \item{트랜스_리놀렌산_18_3t}{해당식품의 트랜스 리놀렌산_18_3t(1회제공량당)}
   \item{식염상당량_g}{해당식품의 식염상당량_g(1회제공량당)}
   \item{회분_g}{해당식품의 회분_g(1회제공량당)}
   \item{폐기율_percent}{해당식품의 폐기율_percent(1회제공량당)}
   \item{성분표출처}{해당 정보의 성분표출처
   농진청, USDA, 식약처, 축과원, 축산연 등}
   \item{발행기관}{해당 정보 발행기관
   국립농업과학원}
}
}
\source{
\url{https://www.data.go.kr/data/15047698/fileData.do}
}
\usage{
foodNutrients
}
\description{
식품의약품안전처에서 제공하는 농축수산물, 가공식품 및 음식 영양성분 데이터베이스. 이용허락범위: 저작자표시(CC BY). 
본 저작물은 '식품의약품안전처'에서 '2020년'작성하여 공공누리 제1유형으로 개방한 '식품영양성분 데이터베이스를'을 이용하였으며, 
해당 저작물은 '식품의약품안전처 식품영양성분 데이터베이스(https://www.foodsafetykorea.go.kr/fcdb/index.do)'에서 무료로 다운받으실 수 있습니다.
}
\examples{
DT::datatable(head(foodNutrients,50))
}
\keyword{datasets}
