% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalBusiness}
\alias{globalBusiness}
\title{대한무역투자진흥공사 해외진출기업 정보}
\format{
A data frame with 19 variables.
\describe{
   \item{구분}{해당 해외진출기업의 대륙 : 아시아}
   \item{국가명}{해당 기업이 진출한 국가명}
   \item{진출지역}{해당 기업이 진출한 지역명}
   \item{회사명.국문.}{해당 해외진출기업 국문 회사명}
   \item{회사명.영문.}{해당 해외진출기업 영문 회사명}
   \item{주소}{해당 해외진출기업 주소}
   \item{홈페이지}{해당 해외진출기업 홈페이지}
   \item{진출년도}{해당 해외진출기업 진출년도}
   \item{진출형태}{해당 해외진출기업 진출형태 : 생산법인, 지점, 판매법인, 서비스법인,연락사무소}
   \item{투자형태}{해당 해외진출기업 투자형태 : M&A, 단독, 합작}
   \item{외국사.합작지분}{해당 해외진출기업 외국사 합작지분}
   \item{업종1}{해당 해외진출기업 업종 : 제조업, 금융·보험업, 도매 및 소매업, 운수업, 농업·임업 및 어업, 건설·공사업, 서비스업, 광업·자원개발, 부동산 및 임대업, 기타, 제조업}
   \item{업종2}{해당 해외진출기업 업종 : 목재·가구·펄프·종이, 식료품, 출판·영상·방송통신 및 정보, 사업 시설관리 및 사업지원, 예술·스포츠 및 여가 관련, 기타 서비스업 등}
   \item{취급분야}{해당 해외진출기업 취급분야 : 생산법인, 지점, 판매법인, 서비스법인,연락사무소}
   \item{본사파견}{해당 해외진출기업 본사파견수}
   \item{현지채용}{해당 해외진출기업 현지채용수}
   \item{내수.수출}{해당 해외진출기업 내수 및 수출 상태 : 수출 및 내수 병행, 100\% 주재국 내수시장 공급, 100\% 제3국 수출}
   \item{모기업명}{해당 해외진출기업 모기업명}
   \item{관할무역관}{해당 해외진출기업 관할무역관명 : 오클랜드무역관, 비엔티안무역관, 쿠알라룸푸르무역관, 양곤무역관, 호치민무역관, 하노이무역관}
 }
}
\source{
\url{https://www.data.go.kr/data/15003297/fileData.do}
}
\usage{
globalBusiness
}
\description{
대한무역투자진흥공사에서 제공하는 전세계 124개국의 해외진출기업의 회사명, 주소, 홈페이지, 진출년도, 진출형태, 투자형태 등의 정보.
}
\examples{
DT::datatable(head(globalBusiness,50))
}
\keyword{datasets}
