% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gyeonggiER}
\alias{gyeonggiER}
\title{경기도 응급의료기관 및 응급의료지원센터 현황}
\format{
A data frame with 14 variables.
}
\usage{
gyeonggiER
}
\description{
경기도 내 응급의료기관 및 응급의료지원센터 현황. 기관 및 센터의 소재지 주소와 대표 전화번호 등의 정보를 제공. 이용허락범위 제한 없음.
Last Update: 2022-05-02;
https://www.data.go.kr/data/15011313/fileData.do
}
\examples{
library(leaflet)

gyeonggiER |> 
leaflet() |> 
  addTiles() |> 
  addMarkers(~경도, ~위도, popup = ~as.character(`병원명/센터명`), label = ~as.character(`병원명/센터명`)) |> 
  addMarkers(~병원경도, ~병원위도, popup = ~as.character(기관명), label = ~as.character(기관명), data = datatoys::seoulER)
}
\keyword{datasets}
