% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatching}
\alias{hatching}
\title{행정안전부 부화업 정보}
\format{
A data frame with 9 variables.
\describe{
   \item{사업장명}{해당 부화업 사업장명}
   \item{전화번호}{해당 부화업 전화번호}
   \item{소재지면적}{해당 부화업 소재지면적}
   \item{소재지우편번호}{해당 부화업 소재지우편번호}
   \item{소재지전체주소}{해당 부화업 소재지전체주소}
   \item{도로명전체주소}{해당 부화업 도로명전체주소}
   \item{도로명우편번호}{해당 부화업 도로명우편번호}
   \item{경도}{해당 부화업 경도}
   \item{위도}{해당 부화업 위도}
 }
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
hatching
}
\description{
닭 또는 오리 등 부화대상 가축의 알을 인공부화시설로 부화시켜 판매하는 업소정보, 주소, 번호, 경도, 위도 등을 제공
}
\examples{
library(leaflet)
leaflet(hatching) \%>\% 
  addTiles() \%>\% 
  setView(lng=127.9784, lat=35.866, zoom=7) \%>\%  
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircleMarkers(lng=~`경도`, lat=  ~`위도`,
                   label =~`사업장명`,
                   popup =~paste(`사업장명`,"<BR>", "주소 : ",`도로명전체주소`, sep="","<BR>","전화번호 : ",`전화번호`))
}
\keyword{datasets}
