% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitalInfo}
\alias{hospitalInfo}
\title{hospitalInfo}
\format{
A data frame with 29 variables.
}
\usage{
hospitalInfo
}
\description{
건강보험심사평가원에서 제공하는 병의원 기본정보 데이터셋.
Last Update: 2022-04-14;
https://www.data.go.kr/data/15051059/fileData.do
}
\keyword{datasets}
