% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housingPrice}
\alias{housingPrice}
\title{housingPrice}
\format{
A data frame with 20 variables.
}
\usage{
housingPrice
}
\description{
국토교통부에서 제공하는 2021년도 공동주택 공시가격정보입니다. 
Last Update: 2021-10-29;
https://www.data.go.kr/data/3073746/fileData.do
}
\keyword{datasets}
