% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iceMarket}
\alias{iceMarket}
\title{행정안전부 식용얼음판매업 정보}
\format{
A data frame with 9 variables:
\describe{
  \item{사업장명}{해당 식용얼음판매업 사업장명}
  \item{전화번호}{해당 식용얼음판매업 전화번호}
  \item{소재지면적}{해당 식용얼음판매업 소재지면적}
  \item{소재지우편번호}{해당 식용얼음판매업 소재지우편번호}
  \item{소재지전체주소}{해당 식용얼음판매업 소재지전체주소}
  \item{도로명전체주소}{해당 식용얼음판매업 도로명전체주소}
  \item{도로명우편번호}{해당 식용얼음판매업 도로명우편번호}
  \item{경도}{해당 식용얼음판매업 경도}
  \item{위도}{해당 식용얼음판매업 위도}
}
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
iceMarket
}
\description{
식용얼음을 전문적으로 판매하는 업소정보	, 주소, 번호, 경도, 위도 등을 제공
}
\examples{
library(leaflet)
leaflet(iceMarket) \%>\% 
  addTiles() \%>\% 
  setView(lng=128.9784, lat=35.266, zoom=9) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~`경도`, lat=  ~`위도`,
             label =~`사업장명`,
             popup =~paste(`사업장명`,"<BR>", "주소 : ",`도로명전체주소`, sep="","<BR>","전화번호 : ",`전화번호`))
}
\keyword{datasets}
