% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imdependenceMov}
\alias{imdependenceMov}
\title{공공데이터포털 독립운동정보 원문정보}
\format{
A data frame with 3 variables.
\describe{
   \item{순번}{독립운동정보 원문자료 순번}
   \item{자료번호}{독립운동정보 원문자료 자료번호}
   \item{내용}{독립운동정보 원문자료 내용}
 }
}
\source{
\url{https://www.data.go.kr/data/15070349/fileData.do?recommendDataYn=Y#/tab-layer-recommend-data}
}
\usage{
imdependenceMov
}
\description{
독립기념관 독립운동정보 원문자료의 자료번호, 자료내용 등의 정보입니다.
}
\examples{
library(tidytext)
library(ggwordcloud)
library(stringr)
library(ggthemes)
imdependenceMov$내용 <- imdependenceMov$내용 \%>\% 
  str_replace_all('[^가-힣]', ' ') \%>\% 
  str_squish()

# mask_png <- png::readPNG("korean-flag-1424985_1280.png")

imdependenceMov \%>\% 
  unnest_tokens(input = 내용, output = word, token = "words") \%>\% 
  count(word, sort = TRUE) \%>\% 
  filter(str_count(word) > 1) \%>\% 
  head(100) \%>\% 
  arrange(-n) \%>\% 
  ggplot(aes(label = word, 
             size = n, 
             col = as.character(n))) +
  geom_text_wordcloud(seed = 1234,
                      max_steps = 1,
                      grid_size = 1,
                      eccentricity = .9,
                      family = "NanumGothic") +
  scale_size_area(max_size=12) +
  theme_minimal()
}
\keyword{datasets}
