% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobMating}
\alias{jobMating}
\title{행정안전부 직업소개소 정보}
\format{
A data frame with 9 variables.
\describe{
   \item{사업장명}{해당 직업소개소 사업장명}
   \item{전화번호}{해당 직업소개소 전화번호}
   \item{소재지우편번호}{해당 직업소개소 소재지우편번호}
   \item{소재지전체주소}{해당 직업소개소 소재지전체주소}
   \item{도로명전체주소}{해당 직업소개소 도로명전체주소}
   \item{도로명우편번호}{해당 직업소개소 도로명우편번호}
   \item{경도}{해당 직업소개소 경도}
   \item{위도}{해당 직업소개소 위도}
   \item{유무료}{해당 직업소개소 유무료 여부 : 유료, 무료}
 }
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
jobMating
}
\description{
유료/무료로 구인 또는 구직의 신청을 받아 구인자와 구직자 간에 고용계약이 성립되도록 알선하는 업소정보, 주소, 번호, 경도, 위도 등을 제공
}
\examples{
free_color = colorFactor('Set1', jobMating$유무료)

library(leaflet)
leaflet(jobMating) \%>\% 
  addTiles() \%>\% 
  setView(lng=127.9784, lat=36.866, zoom=7) \%>\%  
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~`경도`, lat=  ~`위도`,
             label =~`사업장명`,
             popup =~paste(`사업장명`,"<BR>", "주소 : ",`도로명전체주소`, sep="","<BR>","전화번호 : ",`전화번호`),
             color = ~free_color(유무료)) \%>\% 
  addLegend(position = 'bottomright', 
            title = '유무료', 
            pal = free_color, values = ~유무료, opacity = 0.5)
}
\keyword{datasets}
