% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{karaoke}
\alias{karaoke}
\title{행정안전부 단란주점 영업 정보}
\format{
A list with 48 data frame.
\describe{
   \item{번호}{번호}
   \item{개방서비스명}{단란주점영업}
   \item{개방서비스아이디}{개방서비스아이디}
   \item{개방자치단체코드}{개방자치단체코드}
   \item{관리번호}{관리번호}
   \item{인허가일자}{인허가일자}
   \item{인허가취소일자}{인허가취소일자}
   \item{영업상태구분코드}{영업상태구분코드}
   \item{영업상태명}{영업/정상}
   \item{상세영업상태코드}{상세영업상태코드}
   \item{상세영업상태명}{영업}
   \item{폐업일자}{폐업일자}
   \item{휴업시작일자}{휴업시작일자}
   \item{휴업종료일자}{휴업종료일자}
   \item{재개업일자}{재개업일자}
   \item{소재지전화}{소재지전화}
   \item{소재지면적}{소재지면적}
   \item{소재지우편번호}{소재지우편번호}
   \item{소재지전체주소}{소재지전체주소}
   \item{도로명전체주소}{도로명전체주소}
   \item{도로명우편번호}{도로명우편번호}
   \item{사업장명}{사업장명}
   \item{최종수정시점}{최종수정시점}
   \item{데이터갱신구분}{데이터갱신구분}
   \item{데이터갱신일자}{데이터갱신일자}
   \item{업태구분명}{단란주점}
   \item{좌표정보.x.}{좌표정보.x.}
   \item{좌표정보.y.}{좌표정보.y.}
   \item{위생업태명}{단란주점}
   \item{남성종사자수}{남성종사자수}
   \item{여성종사자수}{여성종사자수}
   \item{아파트지역, 기타, 유흥업소밀집지역, 결혼예식장주변, 학교정화(상대), 주택가주변, 학교정화(절대)}{영업장주변구분명}
   \item{등급구분명}{우수, 기타, 자율, 지도, 관리, 을, 갑}
   \item{급수시설구분명}{상수도전용, 지하수전용, 간이상수도, 상수도(음용)지하수(주방용)겸용, 전용상수도(특정시설의 자가용 수도)}
   \item{총종업원수}{총종업원수}
   \item{본사종업원수}{본사종업원수}
   \item{공장사무직종업원수}{공장사무직종업원수}
   \item{공장판매직종업원수}{공장판매직종업원수}
   \item{공장생산직종업원수}{공장생산직종업원수}
   \item{건물소유구분명}{건물소유구분명}
   \item{보증액}{보증액}
   \item{월세액}{월세액}
   \item{다중이용업소여부}{Y, N}
   \item{시설총규모}{시설총규모}
   \item{전통업소지정번호}{전통업소지정번호}
   \item{전통업소주된음식}{전통업소주된음식}
   \item{홈페이지}{홈페이지}
 }
}
\usage{
karaoke
}
\description{
행정안전부에서 제공하는 단란주점 영업 정보. 주로 주류와 함께 음식류를 조리 및 판매하는 곳으로 손님이 노래를 부르는 행위가 허용되는 업소정보 데이터로 인허가일자, 영업상태, 사업장명, 소재지주소 등의 정보등을 제공. * 좌표계 : 중부원점TM(EPSG:2097).
Last Update: 2021-11-19;
https://www.data.go.kr/data/15045017/fileData.do
}
\keyword{datasets}
