% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kcalendar}
\alias{Kcalendar}
\title{한국천문연구원 특일 정보}
\format{
A dataframe with 11 variables.
\describe{
   \item{양력날짜}{양력 날짜로 태양의 자전과 공전 주기를 기준으로 연도를 계산}
   \item{음력날짜}{음력 날짜로 달의 위상과 월의 주기를 기준으로 연도를 계산}
   \item{음력간지년한글}{음력간지년(한글)}
   \item{음력간지월한글}{음력간지월(한글)}
   \item{음력간지일한글}{음력간지일(한글)}
   \item{음력간지년한자}{음력간지년(한자)}
   \item{음력간지월한자}{음력간지월(한자)}
   \item{음력간지일한자}{음력간지일(한자)}
   \item{율리우스적일}{율리우스력의 기원전 4713년 1월 1일 월요일 정오(세계표준시 기준)를 기점으로 계산한 날짜}
   \item{손없는날}{손없는날의 논리값(TRUE or FALSE)}
   \item{손있는날}{손있는날의 방향(동, 서, 남, 북)}
 }
}
\source{
\url{https://www.data.go.kr/data/15012690/openapi.do}
}
\usage{
Kcalendar
}
\description{
한국천문연구원에서 제공하는 국경일정보, 공휴일정보, 기념일정보, 24절기정보, 잡절정보
날짜, 순번, 특일정보의 분류, 공공기관 휴일 여부, 명칭 제공. 이용허락범위 제한없음.
}
\examples{
DT::datatable(head(Kcalendar))
}
\keyword{datasets}
