% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicalCheckup}
\alias{medicalCheckup}
\title{국민건강보험 일반건강검진결과}
\format{
A data frame with 31 variables.
\describe{
 \item{기준년도}{기준년도}
 \item{가입자.일련번호}{가입자.일련번호}
 \item{시도코드}{시도코드}
 \item{성별코드}{성별코드}
 \item{연령대.코드.5세단위.}{연령대 코드_5세단위}
 \item{신장.5Cm단위.}{신장_5Cm 단위}
 \item{체중.5Kg.단위.}{체중_5Kg 단위.}
 \item{허리둘레}{허리둘레}
 \item{시력.좌.}{시력_좌}
 \item{시력.우.}{시력_우}
 \item{청력.좌.}{청력_좌}
 \item{청력.우.}{청력_우}
 \item{수축기.혈압}{수축기 혈압}
 \item{이완기.혈압}{이완기 혈압}
 \item{식전혈당.공복혈당.}{식전혈당 공복혈당}
 \item{총.콜레스테롤}{총 콜레스테롤}
 \item{트리글리세라이드}{트리글리세라이드}
 \item{HDL.콜레스테롤}{HDL 콜레스테롤}
 \item{LDL.콜레스테롤}{LDL 콜레스테롤}
 \item{혈색소}{혈색소}
 \item{요단백}{요단백}
 \item{혈청크레아티닌}{혈청크레아티닌}
 \item{X.혈청지오티.AST}{X.혈청지오티.AST}
 \item{X.혈청지오티.ALT}{X.혈청지오티.ALT}
 \item{감마.지티피}{감마.지티피}
 \item{흡연상태}{흡연상태}
 \item{음주여부}{음주여부}
 \item{구강검진.수검여부}{구강검진 수검여부}
 \item{치아우식증유무}{치아우식증유무}
 \item{치석}{치석}
 \item{데이터.공개일자}{데이터 공개일자}
 }
}
\usage{
medicalCheckup
}
\description{
국민건강보험의 직장가입자와 40세 이상의 피부양자, 세대주인 지역가입자와 40세 이상의 지역가입자의 일반건강검진 결과와 이들 일반건강검진 
대상자 중에 만40세와 만66세에 도달한 이들이 받게 되는 생애전환기건강진단 수검이력이 있는 각 연도별 수진자 100만 명에 대한 
기본정보(성, 연령대, 시도코드 등)와 검진내역(신장, 체중, 총콜레스테롤, 혈색소 등)으로 구성된 개방데이터
https://www.data.go.kr/data/15007122/fileData.do
}
\keyword{datasets}
