% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nationalPension}
\alias{nationalPension}
\title{국민연금공단 국민연금사업장 정보}
\format{
A data frame with 22 variables.
}
\usage{
nationalPension
}
\description{
국민연금공단에서 제공하는 법정동단위 지역별, 국민연금 가입 사업장 정보(단, 개인사업장 및 2인 이하 법인 사업장 정보 미제공).
Last Update: 2022-05-18;
https://www.data.go.kr/data/15083277/fileData.do
}
\keyword{datasets}
