% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openData}
\alias{openData}
\title{공공데이터활용지원센터 공공데이터포털 목록개방현황}
\format{
A list with 13 data frame.
}
\usage{
openData
}
\description{
공공기관이 등록하여 공공데이터포털에서 개방중인 목록 정보(목록명, 목록타입(파일, API), 표준데이터 여부, 국가중점여부, 등록기관, 
기관 분류, 분류체계, 등록일, 마지막 수정일) 를 제공. 이용허락범위 제한 없음.
Last Update: 2022-07-01;
https://www.data.go.kr/data/15062804/fileData.do
}
\keyword{datasets}
