% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optician}
\alias{optician}
\title{행정안전부 안경업 정보}
\format{
A data frame with 10 variables:
\describe{
  \item{사업장명}{해당 안경업 사업장명}
  \item{영업상태명}{해당 안경업 영업상태명 : 취소/말소/만료/정지/중지, 영업/정상,폐업,휴업}
  \item{폐업일자}{해당 안경업 폐업일자}
  \item{전화번호}{해당 안경업 전화번호}
  \item{소재지우편번호}{해당 안경업 소재지우편번호}
  \item{소재지전체주소}{해당 안경업 소재지전체주소}
  \item{도로명전체주소}{해당 안경업 도로명전체주소}
  \item{도로명우편번호}{해당 안경업 도로명우편번호}
  \item{경도}{해당 안경업 경도}
  \item{위도}{해당 안경업 위도}
}
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
optician
}
\description{
눈을 보호하거나 시력의 교정을 요하는 대상에게 시력의 교정 및 안경을 판매 또는 수리하는 업소정보, 주소, 번호, 경도, 위도 등을 제공
}
\examples{
library(leaflet)
leaflet(optician) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=8) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~`경도`, lat=~`위도`,
             label =~`사업장명`,
             popup =~paste(`사업장명`,"<BR>", "주소 : ",`도로명전체주소`, sep="","<BR>","전화번호 : ",`전화번호`))
}
\keyword{datasets}
