% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{국립환경과학원 축산오염원조사정보}
\format{
A data frame with 18 variables
\describe{
  \item{법정동코드}{해당축산계 오염원 법정동코드
  법정동 : 법률로 지정한 행정구역 단위
  법정동코드 : 10자리로 구성, 광역시(2자리) + 시군구(3자리) + 읍면동(3자리) + 리(2자리)
  (예) 서울특별시(11) + 용산구(170) + 한남동(131) + 리(00) = 1117013100}
  \item{시도}{해당 축산계 오염원 시 또는 도}
  \item{시군구}{해당 축산계 오염원 시 또는 군, 구}
  \item{읍면동}{해당 축산계 오염원 읍 또는 면, 동}
  \item{리}{해당축산계 오염원 리}
  \item{축종}{해당 축산계 오염원의 축종
  사슴, 마필(말), 산양(염소포함), 유우(젖소), 개, 한우(소), 돼지, 닭, 가금기타, 오리, 면양(육양포함), 타조}
  \item{법적규제}{법적규제
  신고, 허가, 신고미만}
  \item{사육두수.두.}{해당 축산계 오염원의 사육두수}
  \item{농가수.호.}{해당 축산계 오염원의 농가수}
  \item{가축분뇨발생량.m3.일.}{해당 축산계 오염원에서 발생하는 가축분뇨발생량(1일 기준)}
  \item{자가처리_정화방류.m3.일.}{자가처리 중 정화방류양}
  \item{자가처리_퇴비화.m3.일.}{자가처리 중 퇴비화양(1일기준)
  퇴비화 : 축산농가에서 가축분뇨를 발효시켜 만든 비료성분이 있는 물질 중 액비를 제외한 물질로서 농림축산식품부령으로 정하는 기준에 적합한 것을 생산하는 경우}
  \item{자가처리_액비화.m3.일.}{자가처리 중 액비화양(1일 기준)
  축산농가에서 가축분뇨를 액체 상태로 발효시켜 만든 비료성분이 있는 물질로서 농림축산식품부령으로 정하는 기준에 적합한 것을 생산하는 경우}
  \item{미처리.m3.일.}{해당 축산계 오염원에서 발생하는 미처리양(1일)
  미처리 :  가축 방목·방사로 발생된 분뇨를 상기 처리방법 이외 단순 퇴적, 무단방류 등을 하는 경우}
  \item{위탁처리_공공처리시설.m3.일.}{위탁처리 중 공공처리시설양(1일기준)
  공공처리시설(지자체 운영) : 가축분뇨법 제2조 9항의 따라 지방자치단체의 장이 설치하는 처리시설과「농업협동조합법」 제2조에 따른 조합 및 중앙회(이하 "농협조합"이라 한다)가 제24조제3항에 따라 특별시장·광역시장·도지사(이하 "시·도지사"라 한다), 특별자치시장 또는 특별자치도지사의 승인을 받아 설치하는 자원화시설로 이송하여 처리하는 경우}
  \item{위탁처리_공동처리시설.m3.일.}{위탁처리 중 공동처리시설양(1일기준)
  공동처리시설(개인 또는 법인) : 가축분뇨법 제12조에 의해  배출시설이 연접하여 위치한 경우(같은 시군구에 위치하지 않은 경우 포함)로서 공동으로 자원화시설 또는 정화시설을 설치하여 처리하는 경우} 
  \item{위탁처리_재활용신고자.m3.일.}{위탁처리 중 재활용신고자양(1일기준)
   재활용신고자(개인 또는 법인) : 가축분뇨법 제27조 1항에 따라 환경부령으로 정하는 양 이상의 가축분뇨를 재활용(퇴비 또는 액비로 만드는 것에 한정한다.)하거나 재활용을 목적으로 가축분뇨를 수집·운반하기 위하여 신고 받은 자 또는 제11조1항, 제3항에 따라 설치허가·신고를 받은 자, 제28조 제1항, 제2호의 가축분뇨처리업의 허가를 받은 자가 가축분뇨를 재활용하는 경우에 해당하는 시설로 이송하여 처리하는 경우}
  \item{위탁처리_가축분뇨처리업.m3.일.}{위탁처리 중 가축분뇨처리업양(1일기준)
  가축분뇨처리업(개인 또는 법인) : 가축분뇨법 제28조에 따른 가축분뇨관련영업 중 자원화시설(퇴비·액비를 만드는 시설은 제외한다.) 또는 정화시설을 갖추어 가축분뇨를 최종적으로 안전하게 처리하는 시설로 이송하여 처리하는 경우}
 }
}
\source{
\url{https://www.data.go.kr/data/3045217/fileData.do}{공공데이터포털:환경 - 상하수도·수질}
}
\usage{
pollution
}
\description{
국립환경과학원에서 제공하는 축산계 오염원에 대한 조사정보
}
\examples{
DT::datatable(head(pollution, 50))
library(highcharter)
library(dplyr)
library(datatoys)

pollution |> 
  group_by(축종) |> 
  summarise(
    총분뇨발생량 = sum(가축분뇨발생량.m3.일.)
  ) |> 
  ungroup() |> 
  arrange(desc(총분뇨발생량)) |> 
  hchart(type = "column", hcaes(축종, 총분뇨발생량), name = "총발생량") |> 
  hc_title(
    text = "축종별 가축분뇨 총 발생량",
    margin = 20,
    align = "left"
  ) |> 
  hc_add_theme(hc_theme_538())
}
\keyword{datasets}
