% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{통계청 인구총조사}
\format{
A data frame with 5 variables:
\describe{
  \item{행정구역별}{해당 연도 대한민국 영토 내에 상주하는 모든 내국인·외국인과 이들이 살고 있는 거처}
  \item{연령별}{해당 연도에 조사된 인구를 연령별(5살)로 나눈 연령대}
  \item{연도}{조사기준 시점 현재}
  \item{분류}{각 행정구역 연령별 인구수를 성별로 나눈 기준
  남자, 여자}
  \item{인구수}{해당 연도, 각 행정구역 연령별 인구 수}
}
}
\source{
\url{https://kosis.kr/statHtml/statHtml.do?orgId=101&tblId=DT_1IN0001_ENG&conn_path=I2}
}
\usage{
population
}
\description{
통계청에서 제공하는 인구총조사,  대한민국 영토 내에 상주하는 모든 내국인·외국인
}
\examples{
DT::datatable(head(population, 50))
}
\keyword{datasets}
