% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postnatal}
\alias{postnatal}
\title{행정안전부 산후조리업 정보}
\format{
A data frame with 27 variables:
\describe{
  \item{사업장명}{해당 산후조리업 사업장명}
  \item{전화번호}{해당 산후조리업 전화번호}
  \item{소재지우편번호}{해당 산후조리업 소재지우편번호}
  \item{소재지전체주소}{해당 산후조리업 소재지전체주소}
  \item{도로명전체주소}{해당 산후조리업 도로명전체주소}
  \item{도로명우편번호}{해당 산후조리업 도로명우편번호}
  \item{경도}{해당 산후조리업 경도}
  \item{위도}{해당 산후조리업 위도}
  \item{임산부정원수}{해당 산후조리업 임산부정원수}
  \item{영유아정원수}{해당 산후조리업 영유아정원수}
  \item{임산부실면적}{해당 산후조리업 임산부실면적}
  \item{영유아실면적}{해당 산후조리업 영유아실면적}
  \item{모유수유실면적}{해당 산후조리업 모유수유실면적}
  \item{급식시설면적}{해당 산후조리업 급식시설면적}
  \item{세탁실면적}{해당 산후조리업 세탁실면적}
  \item{목욕실면적}{해당 산후조리업 목욕실면적}
  \item{조리원화장실면적}{해당 산후조리업 조리원화장실면적}
  \item{사무실면적}{해당 산후조리업 사무실면적}
  \item{간호사수}{해당 산후조리업 간호사수}
  \item{간호조무사수}{해당 산후조리업 간호조무사수}
  \item{영양사수}{해당 산후조리업 영양사수}
  \item{취사부수}{해당 산후조리업 취사부수}
  \item{미화원수}{해당 산후조리업 미화원수}
  \item{기타인원수}{해당 산후조리업 기타인원수}
  \item{건물층수}{해당 산후조리업 건물층수}
  \item{지상층수}{해당 산후조리업 지상층수}
  \item{지하층수}{해당 산후조리업 지하층수}
}
}
\source{
\url{https://www.localdata.go.kr/devcenter/dataDown.do?menuNo=20001}
}
\usage{
postnatal
}
\description{
출산 이후 임산부 및 신생아의 건강과 위생을 관리하기 위한 시설 정보, 주소, 번호, 경도, 위도, 임산부정원수, 영유아정원수 등을 제공
}
\examples{
library(leaflet)
leaflet(postnatal) \%>\% 
  addTiles() \%>\% 
  setView(lng=127.9784, lat=36.866, zoom=7) \%>\%  
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~`경도`, lat=  ~`위도`,
             label =~`사업장명`,
             popup =~paste(`사업장명`,"<BR>", "주소 : ",`도로명전체주소`, sep="","<BR>","전화번호 : ",`전화번호`))
}
\keyword{datasets}
