% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{restaurant}
\alias{restaurant}
\title{경기도 맛집 정보}
\format{
A data frame with 9 variables.
\describe{
   \item{시군명}{시군명}
   \item{음식점명}{음식점명}
   \item{맛집전화번호 }{맛집전화번호}
   \item{대표음식명 }{대표음식명}
   \item{소재지우편번호}{소재지우편번호}
   \item{소재지도로명주소}{소재지도로명주소}
   \item{소재지지번주소}{소재지지번주소}
   \item{WGS84위도}{WGS84위도}
   \item{WGS84경도}{WGS84경도}
 }
}
\usage{
restaurant
}
\description{
경기도 지역 특산물을 활용한 향토 특색음식점으로 경기도를 대표할 수 있는 맛, 위생, 서비스가 우수한 업소를 발굴하여 
경기으뜸맛집으로 육성하고자 선정된 맛집 현황입니다. 음식점명, 맛집전화번호, 대표음식명, 주소, 위경도 등의 정보를 제공합니다.
Last Update: 2017-11-20;
https://www.data.go.kr/data/15057234/openapi.do
}
\keyword{datasets}
