% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scholarship}
\alias{scholarship}
\title{한국장학재단 2020년도 장학금 수혜현황}
\format{
A data frame with 16 variables.
\describe{
  \item{대학명}{대학명}
  \item{학제별}{전문대학, 대학}
  \item{설립별}{사립, 국공립}
  \item{지역별}{지역별}
  \item{교외장학금.소계.원.}{교외장학금 소계.원}
  \item{교외장학금.국가}{교외장학금 국가}
  \item{교외장학금.지방자치단체}{교외장학금 지방자치단체}
  \item{교외장학금.사설및기타}{교외장학금 사설및기타}
  \item{교내장학금.소계.원.}{교내장학금 소계.원}
  \item{교내장학금.성적우수장학금}{교내장학금 성적우수장학금}
  \item{교내장학금.저소득층장학금}{교내장학금 저소득층장학금}
  \item{교내장학금.근로장학금}{교내장학금 근로장학금}
  \item{교내장학금.재난장학금}{교내장학금 재난장학금}
  \item{교내장학금.교직원장학금"}{교내장학금 교직원장학금}
  \item{교내장학금.기타}{교내장학금 기타}
  \item{총계.원.}{총계.원}
  }
}
\usage{
scholarship
}
\description{
한국장학재단에서 제공하는 전년도 장학금 수혜현황(대학정보공시 항목) 조사를 통해, 대학별 재원구분에 따른 교외장학금 수혜 금액 및 
성격유형에 따른 교내장학금 수혜금액 제공. 2020 회계연도부터 재난을 사유로 지급한 교내장학금 추가 조사.
https://www.data.go.kr/data/15038576/fileData.do
}
\keyword{datasets}
