% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulCarpark}
\alias{seoulCarpark}
\title{서울특별시 공영주차장 안내 정보}
\format{
A data frame with 19 variables:
\describe{
  \item{주차장코드}{해당 주차장 주차장코드}
  \item{주차장명}{해당 주차장 주차장명}
  \item{주소}{해당 주차장 주소}
  \item{주차장종류}{해당 주차장 주차장종류 : 노외주차장/ 노상주차장}
  \item{운영구분}{해당 주차장 운영구분 : 시간제 주차장/시간제 + 거주자 주차장/버스전용 주차장/이륜차 전용 주차장/거주자 우선 주차장/시간제 + 버스전용 주차장}
  \item{전화번호}{해당 주차장 전화번호}
  \item{주차현황제공여부}{해당 주차장 주차현황제공여부}
  \item{총 주차 가능 대수}{해당 주차장 총 주차 가능 대수}
  \item{유무료구분}{해당 주차장 유무료구분 : 유료/무료}
  \item{토요일 유무료 구분}{해당 주차장 토요일 유무료 구분}
  \item{공휴일 유무료 구분}{해당 주차장 공휴일 유무료 구분}
  \item{월 정기권 금액}{해당 주차장 월 정기권 금액}
  \item{기본 주차 요금}{해당 주차장 기본 주차 요금}
  \item{기본 주차 단위(분)}{해당 주차장 기본 주차 단위(분)}
  \item{추가 주차 요금}{해당 주차장 추가 주차 요금}
  \item{추가 주차 단위(분)}{해당 주차장 추가 주차 단위(분)}
  \item{일 최대 요금}{해당 주차장 일 최대 요금}
  \item{위도}{해당 주차장 위도}
  \item{경도}{해당 주차장 경도}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-13122/S/1/datasetView.do}
}
\usage{
seoulCarpark
}
\description{
공영 노상, 노외 주차장 운영 정보를 안내하고 있으며, 주차장 여건에 따라 실제 정보와 다를 수 있습니다.
}
\examples{
col_fac = colorFactor('Set1', seoulCarpark$운영구분)

library(leaflet)
leaflet(seoulCarpark) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=11) \%>\%  
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~`경도`, lat=  ~`위도`,
             label =~`주차장명`,
             popup =~paste(`주차장명`,"<BR>", "주소 : ",`주소`, sep="","<BR>","운영구분 : ",`운영구분`,"<BR>", "유무료구분 : ",`유무료구분`),
             color = ~col_fac(운영구분)) \%>\% 
  addLegend(position = 'bottomright', 
            title = '구분', 
            pal = col_fac, values = ~운영구분, opacity = 0.5)
}
\keyword{datasets}
