% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulCivic}
\alias{seoulCivic}
\title{서울특별시 2022년 서울 시민생활 데이터}
\format{
A data frame with 147 variables.
\describe{
   \item{}{}
 }
}
\source{
\url{https://data.seoul.go.kr/dataVisual/seoul/seoulLiving.do}
}
\usage{
seoulCivic
}
\description{
서울시와 SK텔레콤이 공공빅데이터와 통신데이터 가명결합을 통해 추정한 서울 행정동단위 성, 연령별 1인가구와 서울시민의 생활특성 정보. 데이터 명세 및 자세한 정보는 아래 url을 통해 제공. 개인, 민간기업 등이 인공지능 학습, 연구, 기술개발 등에 활용할 수 있으나, 학습데이터에 기반한 인공지능 제품개발, 기술연구 논문 등의 결과물에는 데이터의 출처가 서울시 열린데이터광장 임을 명기 필요.
}
\examples{
dplyr::glimpse(seoulCivic)
}
\keyword{datasets}
