% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulCulture}
\alias{seoulCulture}
\title{서울특별시 문화공간정보}
\format{
A data frame with 12 variables:
\describe{
  \item{분류}{문화시설분류(공연장,미술관,도서관 등등)}
  \item{문화시설명}{해당 문화시설 이름}
  \item{주소}{해당 문화시설 주소}
  \item{위도}{해당 문화시설 위도}
  \item{경도}{해당 문화시설 경도}
  \item{전화번호}{해당 문화시설 전화번호}
  \item{홈페이지}{해당 문화시설 홈페이지}
  \item{관람시간}{해당 문화시설 관람시간}
  \item{휴관일}{해당 문화시설 휴관일}
  \item{개관일자}{해당 문화시설 개관일자}
  \item{객석수}{해당 문화시설 객석수}
  \item{대표이미지}{해당 문화시설 이미지}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-15487/S/1/datasetView.do}
}
\usage{
seoulCulture
}
\description{
서울문화포털에서 제공하는 서울특별시 문화공간정보입니다. 문화시설명, 주소, 전화번호, 홈페이지, 관람시간, 관람료, 휴관일, 객석수, 위치 등의 정보를 제공합니다.
}
\examples{
library(leaflet)
pal <- colorFactor("viridis", seoulCulture$분류)
leaflet(seoulCulture) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=11) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~경도, lat=~위도, color = ~pal(분류))
}
\keyword{datasets}
