% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulER}
\alias{seoulER}
\title{서울특별시 응급실 위치 정보}
\format{
A data frame with 34 variables.
\describe{
   \item{기관ID}{해당 병원 기관ID}
   \item{주소}{해당 병원 주소}
   \item{병원분류}{해당 병원분류
   A : 종합병원, B : 병원}
   \item{병원분류명}{해당 병원분류명
   종합병원, 병원}
   \item{응급의료기관코드}{해당 병원 응급의료기관코드
   G001 : 권역응급의료센터, G006 : 지역응급의료센터, G009 : 응급실운영신고기관, G007 : 지역응급의료센터 , G099 : 응급의료기관 이외}
   \item{응급의료기관코드명}{해당 병원 응급의료기관코드명
   권역응급의료센터, 지역응급의료센터, 응급실운영신고기관, 지역응급의료기관, 응급의료기관 이외}
   \item{응급실운영여부}{해당 병원 응급실운영여부
   1}
   \item{비고}{해당 병원 비고}
   \item{기관설명상세}{해당 병원 기관설명상세}
   \item{간이약도}{해당 병원 간이약도}
   \item{기관명}{해당 병원 기관명}
   \item{대표전화1}{해당 병원 대표전화1}
   \item{응급실전화}{해당 병원 응급실전화}
   \item{진료시간.월요일.C}{진료시간 월요일 종료시간}
   \item{진료시간.화요일.C}{진료시간 화요일 종료시간}
   \item{진료시간.수요일.C}{진료시간 수요일 종료시간}
   \item{진료시간.목요일.C}{진료시간 목요일 종료시간}
   \item{진료시간.금요일.C}{진료시간 금요일 종료시간}
   \item{진료시간.토요일.C}{진료시간 토요일 종료시간}
   \item{진료시간.일요일.C}{진료시간 일요일 종료시간}
   \item{진료시간.공휴일.C}{진료시간 공휴일 종료시간}
   \item{진료시간.월요일.S}{진료시간 월요일 오픈시간}
   \item{진료시간.화요일.S}{진료시간 화요일 오픈시간}
   \item{진료시간.수요일.S}{진료시간 수요일 오픈시간}
   \item{진료시간.목요일.S}{진료시간 목요일 오픈시간}
   \item{진료시간.금요일.S}{진료시간 금요일 오픈시간}
   \item{진료시간.토요일.S}{진료시간 토요일 오픈시간}
   \item{진료시간.일요일.S}{진료시간 일요일 오픈시간}
   \item{진료시간.공휴일.S}{진료시간 공휴일 오픈시간}
   \item{우편번호1}{해당 병원 우편번호1}
   \item{우편번호2}{해당 병원 우편번호2}
   \item{병원경도}{해당 병원 병원경도}
   \item{병원위도}{해당 병원 병원위도}
   \item{작업시간}{해당 병원 작업시간}
}
}
\source{
\url{https://www.data.go.kr/data/15088910/fileData.do}{공공데이터포털:보건 - 보건의료}
}
\usage{
seoulER
}
\description{
서울특별시에서 제공하는 서울시 응급실 위치 정보. 주소, 병원분류명, 응급의료기관, 응급실운영여부, 비고, 기관설명상세, 간이약도, 기관명, 대표전화, 응급실 전화, 진료시간, 병원경도, 병원위도 등의 항목으로 구성. 본 저작물은 서울특별시에서 2021년 작성하여 공공누리 제1유형으로 개방한 '서울특별시 응급실 위치 정보(작성자:서울특별시 빅데이터담당관)'을 이용하였으며, 저작물은 '서울시 열린 데이터 광장(http://data.seoul.go.kr/dataList/OA-20338/S/1/datasetView.do)'에서 무료로 다운받으실 수 있습니다.
Last Update: 2021-09-16;
}
\examples{
library(leaflet)

seoulER |> 
leaflet() |> 
  addTiles() |> 
  addMarkers(~병원경도, ~병원위도, popup = ~as.character(기관명), label = ~as.character(기관명))
DT::datatable(head(seoulER,50))
}
\keyword{datasets}
