% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulFestival}
\alias{seoulFestival}
\title{서울특별시 문화행사 정보}
\format{
A data frame with 15 variables:
\describe{
  \item{자치구}{해당 문화행사 주관 자치구}
  \item{공연(행사)명}{해당 문화행사 이름}
  \item{일시}{해당 문화행사 날짜 및 시간}
  \item{장소}{해당 문화행사 장소}
  \item{기관명}{해당 문화행사 주관 기관}
  \item{이용대상}{해당 문화행사 이용대상 : 8세 이상 등}
  \item{이용요금}{해당 문화행사 이용요금}
  \item{출연자정보}{해당 문화행사 출연자 정보}
  \item{프로그램소개}{해당 문화행사 프로그램 소개}
  \item{홈페이지주소}{해당 문화행사 홈페이지 주소}
  \item{대표이미지}{해당 문화행사 대표이미지}
  \item{신청일}{해당 문화행사 신청일}
  \item{위도}{해당 문화행사 장소 위도}
  \item{경도}{해당 문화행사 장소 경도}
  \item{유무료}{해당 문화행사 유무료 구분 : 무료 / 유료}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-15486/S/1/datasetView.do}
}
\usage{
seoulFestival
}
\description{
서울문화포털에서 제공하는 문화행사 정보입니다. 공연, 행사에 대한 장소, 날짜, 기관명, 이용대상, 이용요금, 출연자, 프로그램 등의 정보를 제공합니다.
}
\examples{
library(leaflet)
leaflet(seoulFestival) \%>\% 
addTiles() \%>\% 
setView(lng=126.9784, lat=37.566, zoom=11) \%>\% 
addProviderTiles('CartoDB.Positron') \%>\% 
addCircles(lng=~경도, lat=~위도,
           label =~`공연(행사)명`,
           popup=~paste(`공연(행사)명`,"<BR>", "기간 : ",일시,sep="","<BR>","장소 : ",장소))
}
\keyword{datasets}
