% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulGraduate}
\alias{seoulGraduate}
\title{서울특별시 대학원 통계}
\format{
A list with 4 data frame.
\describe{
   \item{구분_1}{대학원수(개), 학과수(개), 입학정원수, 석사과정학생수, 박사과정학생수, 졸업자현황}
   \item{구분_2}{소계, 석사과정, 박사과정}
   \item{시점}{조사 시점 : 1997 ~ 2022}
   \item{수}{조사 시점의 수(대학원의 수, 학과 수, 입학 정원 수 등)}
 }
}
\source{
\url{https://data.seoul.go.kr/dataList/211/C/2/datasetView.do}
}
\usage{
seoulGraduate
}
\description{
서울열린데이터 광장에서 제공하는 대학원의 학생수(석사ㆍ박사과정), 교원수, 졸업/입학자 현황(석사ㆍ박사과정) 등, 출처 : 한국교육개발원 교육통계연구센터
}
\examples{
seoulDoctor <- seoulGraduate \%>\% 
  filter(구분_1 == "박사과정학생수" | 구분_2 == "박사과정")

ggplot(seoulDoctor, aes(x=시점, y=수, color=구분_1, group = 구분_1)) +
  geom_point() +
  geom_path() +
  scale_y_continuous(labels = scales::comma) +  
  theme_minimal(base_family = "NanumGothic")+
  labs(title="서울시 박사 통계", color = "") + 
  theme(plot.title = element_text(hjust = 0.5, size = 22, color = "royalblue4", face = "bold"), legend.position = "top")
}
\keyword{datasets}
