% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulLibrary}
\alias{seoulLibrary}
\title{서울특별시 공공도서관 현황정보}
\format{
A data frame with 11 variables:
\describe{
  \item{도서관명}{해당 도서관 이름}
  \item{구(코드)}{해당 도서관이 속한 위치한 구의 코드번호}
  \item{구명}{해당 도서관이 위치한 구}
  \item{주소}{해당 도서관의 주소}
  \item{전화번호}{해당 도서관의 전화번호}
  \item{홈페이지(URL)}{해당 도서관의 홈페이지 주소}
  \item{운영시간}{해당 도서관의 운영시간}
  \item{정기휴관일}{해당 도서관의 정기 휴관일}
  \item{도서관구분}{해당 도서관의 구분(공공도서관)}
  \item{위도}{해당 도서관의 위도}
  \item{경도}{해당 도서관의 경도}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-15480/S/1/datasetView.do}
}
\usage{
seoulLibrary
}
\description{
서울특별시 각 자치구의 공공도서관의 현황정보 입니다. 도서관명, 구명, 주소, 전화번호, 홈페이지주소, 운영시간, 정기휴관일, 위치정보 등을 제공합니다.
}
\examples{
library(leaflet)
pal <- colorFactor("viridis", seoulLibrary$`구(코드)`)
leaflet(seoulLibrary) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=11) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~경도, lat=~위도, color = ~pal(`구(코드)`))
}
\keyword{datasets}
