% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulRestroom}
\alias{seoulRestroom}
\title{서울특별시 공중화장실 위치정보}
\format{
A data frame with 4 variables:
\describe{
  \item{주소(구)}{해당 화장실 주소(구)}
  \item{주소(동)}{해당 화장실 주소(동)}
  \item{위도}{해당 화장실 위도}
  \item{경도}{해당 화장실 경도}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-1370/S/1/datasetView.do}
}
\usage{
seoulRestroom
}
\description{
서울시 공중화장실 현황정보를 제공합니다. 좌표정보를 지도위에 표시하여 화장실 위치정보를 제공합니다.
}
\examples{
library(leaflet)
leaflet(seoulRestroom) \%>\% 
  addTiles() \%>\% 
  setView(lng=126.9784, lat=37.566, zoom=11) \%>\% 
  addProviderTiles('CartoDB.Positron') \%>\% 
  addCircles(lng=~경도, lat=~위도,
             label=~`주소(동)`)
}
\keyword{datasets}
