% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoulStatue}
\alias{seoulStatue}
\title{서울특별시 동상 현황}
\format{
A data frame with 7 variables:
\describe{
  \item{작품명}{해당 동상 작품명}
  \item{설치연도}{해당 동상 설치연도}
  \item{작품설치장소}{해당 동상 작품설치장소}
  \item{작품관리기관}{해당 동상 작품관리기관}
  \item{작품주소}{해당 동상 작품주소}
  \item{상세주소}{해당 동상 상세주소}
  \item{작품상세}{해당 동상 작품상세}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-21242/S/1/datasetView.do}
}
\usage{
seoulStatue
}
\description{
서울시에서 관리하는 동상 정보입니다. 작품명, 설치연도, 작품설치장소, 작품관리기관, 작품주소, 상세주소, 작품상세 정보를 제공합니다.
}
\examples{
DT::datatable(head(seoulStatue, 50))
}
\keyword{datasets}
