% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warmingCenter}
\alias{warmingCenter}
\title{서울특별시 한파쉼터현황}
\format{
A data frame with 15 variables:
\describe{
  \item{쉼터명칭}{해당 한파쉼터 명칭}
  \item{상세주소}{해당 한파쉼터 상세주소}
  \item{면적}{해당 한파쉼터 면적}
  \item{이용가능인원}{해당 한파쉼터 이용가능인원}
  \item{난방기보유대수(열풍기)}{해당 한파쉼터 난방기보유대수(열풍기)}
  \item{난방기보유대수(히터)}{해당 한파쉼터 난방기보유대수(히터)}
  \item{난방기보유대수(난로)}{해당 한파쉼터 난방기보유대수(난로)}
  \item{난방기보유대수(라디에이터)}{해당 한파쉼터 난방기보유대수(라디에이터)}
  \item{사용여부}{해당 한파쉼터 사용여부 : Y , N}
  \item{한파쉼터운영시작일}{해당 한파쉼터 운영시작일}
  \item{한파쉼터운영종료일}{해당 한파쉼터 운영종료일}
  \item{한파쉼터평일시작시간}{해당 한파쉼터 평일시작시간}
  \item{한파쉼터평일종료시간}{해당 한파쉼터 평일종료시간}
  \item{한파쉼터주말시작시간}{해당 한파쉼터 주말시작시간}
  \item{한파쉼터주말종료시간}{해당 한파쉼터 주말종료시간}
}
}
\source{
\url{https://data.seoul.go.kr/dataList/OA-21066/S/1/datasetView.do}
}
\usage{
warmingCenter
}
\description{
서울시 한파쉼터 현황입니다. 쉼터별 주소, 면적, 좌표, 인원, 보유품목, 운영시간 등의 정보를 제공합니다.
}
\examples{
DT::datatable(head(warmingCenter, 50))
}
\keyword{datasets}
