% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather2020}
\alias{weather2020}
\title{농림축산식품부 농업 종관기상 데이터}
\format{
A list with 3 data frame.
}
\usage{
weather2020
}
\description{
농림축산식품부에서 제공하는 농업 종관기상 데이터.
Last Update: 2021-03-26;
https://www.data.go.kr/data/15034381/fileData.do
}
\keyword{datasets}
