% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident}
\alias{accident}
\title{도로교통공단 사망교통사고 정보}
\format{
A data frame with 23 variables.
\describe{
  \item{발생년}{발생년}
  \item{발생년월일시}{발생년월일시}
  \item{주야}{주야}
  \item{요일}{요일}
  \item{사망자수}{사망자수}
  \item{부상자수}{부상자수}
  \item{중상자수}{중상자수}
  \item{경상자수}{경상자수}
  \item{부상신고자수}{부상신고자수}
  \item{발생지시도}{발생지시도}
  \item{발생지시군구}{발생지시군구}
  \item{사고유형_대분류}{차대사람, 차대차, 차량단독, 철길건널목}
  \item{사고유형_중분류}{차도통행중, 추돌, 횡단중, 공작물충돌, 도로이탈, 기타, 정면충돌, 측면충돌, 길가장자리구역통행중, 전도, 전복, 보도통행중, 주/정차차량 충돌, 후진중충돌, 철길건널목}
  \item{사고유형}{사고유형}
  \item{가해자법규위반}{안전운전 의무 불이행, 과속, 보행자 보호의무 위반, 중앙선 침범, 신호위반, 기타, 안전거리 미확보, 교차로 통행방법 위반}
  \item{도로형태_대분류}{단일로, 교차로, 기타, 불명, 철길건널목}
  \item{도로형태}{기타단일로, 지하차도(도로)내, 교차로횡단보도내, 교차로내, 교차로부근, 교량위, 기타, 고가도로위, 터널안, 불명, 철길건널}
  \item{가해자_당사자종별}{승용차, 화물차, 농기계, 승합차, 이륜차, 사륜오토바이(ATV), 특수차, 원동기장치자전거, 건설기계, 자전거, 개인형이동수단(PM), 기타}
  \item{피해자_당사자종별}{보행자, 화물차, 없음, 승용차, 농기계, 자전거, 이륜차, 승합차, 건설기계, 원동기장치자전거, 불명, 특수차, 사륜오토바이(ATV), 개인형이동수단(PM), 기타}
  \item{발생위치X_UTMK}{발생위치X_UTMK}
  \item{발생위치Y_UTMK}{발생위치Y_UTMK}
  \item{경도}{경도}
  \item{위도}{위도}
 }
}
\usage{
accident
}
\description{
도로교통공단에서 제공하는 사망 교통사고에 대한 개별정보(발생일시, 사고유형, 위치좌표 등)
부상자수 = 중상자수 + 경상자수 + 부상신고자수
https://www.data.go.kr/data/15070340/fileData.do
}
\examples{
 library(leaflet)
 accident |> 
   dplyr::filter(발생지시도 == "서울") |> 
   dplyr::filter(사망자수 > 0) |> 
   dplyr::filter(피해자_당사자종별 == "보행자") |> 
   leaflet() |> 
   addTiles() |> 
   addMarkers(~경도, ~위도, popup = ~as.character(사고유형), label = ~as.character(사고유형))
}
\keyword{datasets}
