% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident}
\alias{accident}
\title{도로교통공단 사망교통사고 정보}
\format{
A data frame with 23 variables.
\describe{
  \item{발생년}{사망 교통사고가 발생한 연도}
  \item{발생년월일시}{사망 교통사고가 발생한 발생년월일시}
  \item{주야}{사망 교통사고가 발생이 낮에 발생한 것인지, 밤에 발생한 것인지}
  \item{요일}{사망 교통사고가 발생한 요일}
  \item{사망자수}{사망 교통사고로 인한 사망자수
  사망 : 교통사고 발생시로부터 30일이내에 사망한 경우}
  \item{부상자수}{사망 교통사고로 인한 부상자수
  부상자수는 중상자수 + 경상자수 + 부상신고자수이다.}
  \item{중상자수}{사망 교통사고로 인한 중상자수
  중상 : 교통사고로 인하여 3주 이상의 치료를 요하는 부상을 입은 경우)}
  \item{경상자수}{사망 교통사고로 인한 경상자수
  경상 : 교통사고로 인하여 5일 이상 3주 미만의 치료를 요하는 부상을 입은 경우}
  \item{부상신고자수}{사망 교통사고로 인한 부상신고자수
  부상신고 : 교통사고로 인하여 5일 미만의 치료를 요하는 부상을 입은 경우}
  \item{발생지시도}{사망 교통사고가 발생한 시 또는 도}
  \item{발생지시군구}{사망 교통사고로 인한 시 또는 군, 구}
  \item{사고유형_대분류}{해당 사망 교통사고의 유형
  차대사람 : 보행자가 가해운전자 또는 피해운전자가 된 사고, 차대차 :  당해사고의 가해운전자와 피해운전자 모두 차량인 사고, 차량단독 상대방 없이 가해운전자만 있는 경우(추락, 전도, 도로외 이탈 등)와 피해운전자가 주차차량, 물건 및 기타 도로시설물 등인 경우, 철길건널목}
  \item{사고유형_중분류}{해당 사망 교통사고의 유형
  차도통행중, 추돌, 횡단중, 공작물충돌, 도로이탈, 기타, 정면충돌, 측면충돌, 길가장자리구역통행중, 전도, 전복, 보도통행중, 주/정차차량 충돌, 후진중충돌, 철길건널목}
  \item{사고유형}{해당 사망 교통사고의 유형
  차도통행중, 추돌, 횡단중, 공작물충돌, 도로이탈 추락, 도로이탈 기타, 기타, 정면충돌, 측면충돌, 길가장자리구역통행중, 전도, 전복, 보도통행중, 주/정차차량 충돌, 후진중충돌, 철길건널목}
  \item{가해자법규위반}{해당 사망 교통사고 가해자가 위반한 법규사항
  안전운전 의무 불이행, 과속, 보행자 보호의무 위반, 중앙선 침범, 신호위반, 기타, 안전거리 미확보, 교차로 통행방법 위반}
  \item{도로형태_대분류}{해당 사망 교통사고가 일어난 도로의 형태
  단일로, 교차로, 기타, 불명, 철길건널목}
  \item{도로형태}{해당 사망 교통사고가 일어난 도로의 형태
  기타단일로, 지하차도(도로)내, 교차로횡단보도내, 교차로내, 교차로부근, 교량위, 기타, 고가도로위, 터널안, 불명, 철길건널}
  \item{가해자_당사자종별}{해당 사망 교통사고 가해자가 타고 있던 분류
  승용차, 화물차, 농기계, 승합차, 이륜차, 사륜오토바이(ATV), 특수차, 원동기장치자전거, 건설기계, 자전거, 개인형이동수단(PM), 기타}
  \item{피해자_당사자종별}{해당 사망 교통사고 피해자 분류
  보행자, 화물차, 없음, 승용차, 농기계, 자전거, 이륜차, 승합차, 건설기계, 원동기장치자전거, 불명, 특수차, 사륜오토바이(ATV), 개인형이동수단(PM), 기타}
  \item{발생위치X_UTMK}{해당 사망 교통사고가 발생한 위치(X_UTMK)
  UTM 좌표계 : UTM 좌표계란 전 지구상의 점들의 위치를 통일된 체계로 나타내기 위한 격자 좌표체계, UTM-K 좌표계: 한국 지리정보원에 사용하는 좌표로 UTM 좌표계를 한국에 맞게 보정}
  \item{발생위치Y_UTMK}{해당 사망 교통사고가 발생한 위치(Y_UTMK)
  UTM 좌표계 : UTM 좌표계란 전 지구상의 점들의 위치를 통일된 체계로 나타내기 위한 격자 좌표체계
  UTM-K 좌표계: 한국 지리정보원에 사용하는 좌표로 UTM 좌표계를 한국에 맞게 보정}
  \item{경도}{해당 사망 교통사고가 발생한 경도
  경도 : 지구 위의 위치를 나타내는 좌표축 중에서 세로로 된 것}
  \item{위도}{해당 사망 교통사고가 발생한 위도
  지구 위의 위치를 나타내는 좌표축 중에서 가로로 된 것}
 }
}
\source{
\url{https://www.data.go.kr/data/15070340/fileData.do}{공공데이터포털:	공공질서및안전 - 안전관리}
}
\usage{
accident
}
\description{
도로교통공단에서 제공하는 사망 교통사고에 대한 개별정보(발생일시, 사고유형, 위치좌표 등)
부상자수 = 중상자수 + 경상자수 + 부상신고자수
}
\examples{
 library(leaflet)
 accident |> 
   dplyr::filter(발생지시도 == "서울") |> 
   dplyr::filter(사망자수 > 0) |> 
   dplyr::filter(피해자_당사자종별 == "보행자") |> 
   leaflet() |> 
   addTiles() |> 
   addMarkers(~경도, ~위도, popup = ~as.character(사고유형), label = ~as.character(사고유형))
DT::datatable(head(accident, 50))
}
\keyword{datasets}
