% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airport}
\alias{airport}
\title{국토교통부 전세계 공항정보}
\format{
A list with 8 data frame.
}
\usage{
airport
}
\description{
국토교통부에서 제공하는 전세계 공항의 영문 공항명, 한글 공항명, IATA 코드, ICAO 코드, 국가명, 지역명에 대한 정보.
Last Update: 2022-06-10;
https://www.data.go.kr/data/3051587/fileData.do
}
\keyword{datasets}
