% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carInspection}
\alias{carInspection}
\title{한국교통안전공단 자동차검사소 정보}
\format{
A data frame with 22 variables.
\describe{
   \item{자동차검사소명}{해당 자동차검사소명}
   \item{자동차검사소유형}{해당 자동차검사소 유형
   공공}
   \item{소재지도로명주소}{해당 자동차검사소 도로명주소}
   \item{소재지지번주소}{해당 자동차검사소 지번주소}
   \item{위도}{해당 자동차검사소 위도}
   \item{경도}{해당 자동차검사소 경도}
   \item{검사소전화번호}{해당 자동차검사소 전화번호}
   \item{운영시간}{해당 자동차검사소 운영시간}
   \item{검사진로수}{해당 자동차검사소 검사진로수}
   \item{부지면적}{해당 자동차검사소 부지면적}
   \item{건물면적}{해당 자동차검사소 건물면적}
   \item{검사기술인력수}{해당 자동차검사소 검사기술인력수}
   \item{신규검사여부}{해당 자동차검사소 신규검사여부
   Y : 예, N: 아니오}
   \item{정기검사여부}{해당 자동차검사소 정기검사여부
   Y : 예, N: 아니오}
   \item{튜닝검사여부}{해당 자동차검사소 튜닝검사여부
   Y : 예, N: 아니오}
   \item{임시검사여부}{해당 자동차검사소 임시검사여부
   Y : 예, N: 아니오}
   \item{수리검사여부}{해당 자동차검사소 수리검사여부
   Y : 예, N: 아니오}
   \item{배출가스정밀검사여부}{해당 자동차검사소 배출가스정밀검사여부
   Y : 예, N: 아니오}
   \item{택시미터검정여부}{해당 자동차검사소 택시미터검정여부
   Y : 예, N: 아니오}
   \item{관리기관전화번호}{해당 자동차검사소 관리기관전화번호}
   \item{관리기관명}{해당 자동차검사소 관리기관명}
   \item{데이터기준일자}{해당 데이터의 기준일자}
 }
}
\source{
\url{https://www.data.go.kr/data/3041572/fileData.do}
}
\usage{
carInspection
}
\description{
한국교통안전공단에서 운영하고 있는 자동차검사소 정보. 해당 지역별 리스트와 위도, 경도 등 각종 정보를 제공.
}
\examples{
DT::datatable(head(carInspection,50))
}
\keyword{datasets}
