% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{경찰청 범죄 발생 지역별 통계}
\format{
A data frame with 90 variables.
\describe{
  \item{범죄대분류}{강력범죄, 교통범죄, 기타범죄, 노동범죄, 마약범죄, 병역범죄, 보건범죄, 선거범죄, 안보범죄, 절도범죄, 지능범죄, 특별경제범죄, 폭력범죄, 풍속범죄, 환경범죄}
  \item{범죄중분류}{살인기수, 살인미수등 ,강도, 강간, 유사강간, 강제추행, 기타강간강제추행등, 방화, 절도, 상해, 폭행, 체포감금, 협박, 약취유인, 폭력행위등, 공갈, 손괴, 직무유기, 직권남용, 증수뢰, 통화, 문서인장, 유가증권인지, 사기, 횡령, 배임, 성풍속범죄, 도박범죄, 특별경제범죄, 마약범죄, 보건범죄, 환경범죄, 교통범죄, 노동범죄, 안보범죄, 선거범죄, 병역범죄, 기타범죄}
  \item{서울}{서울에서 발생한 범죄 건수}
  \item{부산}{부산에서 발생한 범죄 건수}
  \item{대구}{대구에서 발생한 범죄 건수}
  \item{인천}{인천에서 발생한 범죄 건수}
  \item{광주}{광주에서 발생한 범죄 건수}
  \item{대전}{대전에서 발생한 범죄 건수}
  \item{울산}{울산에서 발생한 범죄 건수}
  \item{세종}{세종에서 발생한 범죄 건수}
  \item{경기.고양}{경기 고양에서 발생한 범죄 건수}
  \item{경기.과천}{경기 과천에서 발생한 범죄 건수}
  \item{경기.광명}{경기 광명에서 발생한 범죄 건수}
  \item{경기.광주}{경기 광주에서 발생한 범죄 건수}
  \item{경기.구리}{경기 구리에서 발생한 범죄 건수}
  \item{경기.군포}{경기 군포에서 발생한 범죄 건수}
  \item{경기.김포}{경기 김포에서 발생한 범죄 건수}
  \item{경기.남양주}{경기 남양주에서 발생한 범죄 건수}
  \item{경기.동두천}{경기 동두천에서 발생한 범죄 건수}
  \item{경기.부천}{경기 부천에서 발생한 범죄 건수}
  \item{경기.성남}{경기 성남에서 발생한 범죄 건수}
  \item{경기.수원}{경기 수원에서 발생한 범죄 건수}
  \item{경기.시흥}{경기 시흥에서 발생한 범죄 건수}
  \item{경기.안산}{경기 안산에서 발생한 범죄 건수}
  \item{경기.안성}{경기 안성에서 발생한 범죄 건수}
  \item{경기.안양}{경기 안양에서 발생한 범죄 건수}
  \item{경기.양주}{경기 양주에서 발생한 범죄 건수}
  \item{경기.여주}{경기 여주에서 발생한 범죄 건수}
  \item{경기.오산}{경기 오산에서 발생한 범죄 건수}
  \item{경기.용인}{경기 용인에서 발생한 범죄 건수}
  \item{경기.의왕}{경기 의왕에서 발생한 범죄 건수}
  \item{경기.의정부}{경기 의정부에서 발생한 범죄 건수}
  \item{경기.이천}{경기 이천에서 발생한 범죄 건수}
  \item{경기.파주}{경기 파주에서 발생한 범죄 건수}
  \item{경기.평택}{경기 평택에서 발생한 범죄 건수}
  \item{경기.포천}{경기 포천에서 발생한 범죄 건수}
  \item{경기.하남}{경기 하남에서 발생한 범죄 건수}
  \item{경기.화성}{경기 화성에서 발생한 범죄 건수}
  \item{강원.강릉}{강원 강릉에서 발생한 범죄 건수}
  \item{강원.동해}{강원 동해에서 발생한 범죄 건수}
  \item{강원.삼척}{강원 삼척에서 발생한 범죄 건수}
  \item{강원.속초}{강원 속초에서 발생한 범죄 건수}
  \item{강원.원주}{강원 원주에서 발생한 범죄 건수}
  \item{강원.춘천}{강원 춘천에서 발생한 범죄 건수}
  \item{강원.태백}{강원 태백에서 발생한 범죄 건수}
  \item{충북.제천}{충북 제천에서 발생한 범죄 건수}
  \item{충북.청주}{충북 청주에서 발생한 범죄 건수}
  \item{충북.충주}{충북 충주에서 발생한 범죄 건수}
  \item{충남.계룡}{충남 계룡에서 발생한 범죄 건수}
  \item{충남.공주}{충남 공주에서 발생한 범죄 건수}
  \item{충남.논산}{충남 논산에서 발생한 범죄 건수}
  \item{충남.당진}{충남 당진에서 발생한 범죄 건수} 
  \item{충남.보령}{충남 보령에서 발생한 범죄 건수} 
  \item{충남.서산}{충남 서산에서 발생한 범죄 건수}
  \item{충남.아산}{충남 아산에서 발생한 범죄 건수}
  \item{충남.천안}{충남 천안에서 발생한 범죄 건수}
  \item{전북.군산}{전북 군산에서 발생한 범죄 건수}
  \item{전북.김제}{전북 김제에서 발생한 범죄 건수}
  \item{전북.남원}{전북 남원에서 발생한 범죄 건수}
  \item{전북.익산}{전북 익산에서 발생한 범죄 건수}
  \item{전북.전주}{전북 전주에서 발생한 범죄 건수}
  \item{전북.정읍}{전북 정읍에서 발생한 범죄 건수}
  \item{전남.광양}{전남 광양에서 발생한 범죄 건수}
  \item{전남.나주}{전남 나주에서 발생한 범죄 건수}
  \item{전남.목포}{전남 목포에서 발생한 범죄 건수}
  \item{전남.순천}{전남 순천에서 발생한 범죄 건수}
  \item{전남.여수}{전남 여수에서 발생한 범죄 건수}
  \item{경북.경산}{경북 경산에서 발생한 범죄 건수}
  \item{경북.경주}{경북 경주에서 발생한 범죄 건수}
  \item{경북.구미}{경북 구미에서 발생한 범죄 건수}
  \item{경북.김천}{경북 김천에서 발생한 범죄 건수}
  \item{경북.문경}{경북 문경에서 발생한 범죄 건수}
  \item{경북.상주}{경북 상주에서 발생한 범죄 건수}
  \item{경북.안동}{경북 안동에서 발생한 범죄 건수}
  \item{경북.영주}{경북 영주에서 발생한 범죄 건수}
  \item{경북.영천}{경북 영천에서 발생한 범죄 건수}
  \item{경북.포항}{경북 포항에서 발생한 범죄 건수}
  \item{경남.거제}{경남 거제에서 발생한 범죄 건수}
  \item{경남.김해}{경남 김해에서 발생한 범죄 건수}
  \item{경남.마산}{경남 마산에서 발생한 범죄 건수}
  \item{경남.밀양}{경남 밀양에서 발생한 범죄 건수}
  \item{경남.사천}{경남 사천에서 발생한 범죄 건수}
  \item{경남.양산}{경남 양산에서 발생한 범죄 건수}
  \item{경남.진주}{경남 진주에서 발생한 범죄 건수}
  \item{경남.창원}{경남 창원에서 발생한 범죄 건수}
  \item{경남.통영}{경남 통영에서 발생한 범죄 건수}
  \item{제주.서귀포}{제주 서귀포에서 발생한 범죄 건수}
  \item{제주.제주}{제주 제주에서 발생한 범죄 건수}
  \item{기타도시}{기타도시에서 발생한 범죄 건수}
  \item{도시이외}{도시이외에서 발생한 범죄 건수}
 }
}
\source{
\url{https://www.data.go.kr/data/3074462/fileData.do}{공공데이터포털:공공질서및안전 - 경찰}
}
\usage{
crime
}
\description{
전국 경찰관서에 고소, 고발, 인지 등으로 형사입건된 사건의 발생, 검거, 피의자에 대한 지역별 분석 현황
범죄대분류, 범죄중분류, 서울, 부산, 대구, 인천, 광주, 대전 울산, 세종, 경기도 지역으로 구분하여 발생 건수 제공
}
\examples{
crime |> 
  tidyr::pivot_longer(
    cols = 서울:도시이외, 
    names_to = "시군", 
    values_to = "건"
  )  
DT::datatable(head(crime, 50))
}
\keyword{datasets}
