% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{경찰청 범죄발생 지역별 통계}
\format{
A data frame with 90 variables.
}
\usage{
crime
}
\description{
전국 경찰관서에 고소, 고발, 인지 등으로 형사입건된 사건의 발생, 검거, 피의자에 대한 지역별 분석 현황
범죄대분류, 범죄중분류, 서울, 부산, 대구, 인천, 광주, 대전 울산, 세종, 경기도 지역으로 구분하여 발생 건수 제공
https://www.data.go.kr/data/3074462/fileData.do
}
\keyword{datasets}
