% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crimePlace}
\alias{crimePlace}
\title{경찰청 범죄 발생 장소별 통계}
\format{
A list with 4 data frame.
\describe{
   \item{범죄대분류}{강력범죄, 절도범죄, 폭력범죄, 지능범죄, 풍속범죄, 특별경제범죄, 마약범죄, 보건범죄, 환경범죄, 교통범죄, 노동범죄, 안보범죄, 선거범죄, 병역범죄, 기타범죄}
   \item{범죄중분류}{살인기수, 살인미수등, 강도, 강간, 유사강간, 강제추행, 기타 강간-강제추행등, 방화, 절도범죄, 상해, 폭행, 체포-감금, 협박, 약취-유인, 폭력행위등, 공갈, 손괴, 직무유기, 직권남용, 증수뢰, 통화, 문서-인장, 유가증권인지, 사기, 횡령, 배임, 성풍속범죄, 도박범죄, 특별경제범죄, 마약범죄, 보건범죄, 환경범죄, 교통범죄, 노동범죄, 안보범죄, 선거범죄, 병역범죄, 기타범죄}
   \item{장소}{아파트.연립다세대, 단독주택, 고속도로, 노상, 백화점, 슈퍼마켓, 편의점, 대형할인매장, 상점, 시장.노점, 숙박업소.목욕탕, 유흥접객업소, 사무실, 공장, 공사장.광산, 창고, 역.대합실, 지하철, 기타표통수단내, 흥행장, 유원지, 학교, 금융기관, 의료기관, 종교기관, 산야, 해상, 부대, 구금장소, 공지, 주차장, 공중화장실, 피씨방}
   \item{발생횟수}{발생횟수}
}
}
\usage{
crimePlace
}
\description{
경찰청에서 제공하는 범죄유형별(강력범죄, 폭력범죄, 지능범죄 등)로 범죄장소(아파트, 단독주택, 고속도로, 노상 ,백화점, 슈퍼마켓 등)별 통계. 이용허락범위 제한 없음.
Last Update: 2022-10-08;
https://www.data.go.kr/data/3074463/fileData.do
}
\keyword{datasets}
