% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elevator}
\alias{elevator}
\title{한국승강기안전공단 국내 승강기 보유 현황}
\format{
A list with 8 data frame.
\describe{
   \item{지역}{해당 승강기 보유 건물의 지역}
   \item{건물명}{해당 승강기 보유 건물명}
   \item{건물층수}{해당 승강기 보유 건물의 건물 층수}
   \item{운행층수}{해당 승강기의 승강기 운행층수}
   \item{승강기종류}{해당 승강기종류
   경사형휠체어리프트, 화물용, 병원용, 승객용, 화물용(DW), 전망용, 덤웨이터, 장애인용, 무빙워크, 장애/승객화물용, 승객화물용, 장애/병원용, 소방구조/장애인용, 자동차용, 장애/전망용, 소방구조/장애/승객화물, 화물/자동차용, 소방구조용, 에스컬레이터 등}
   \item{건물용도_대}{해당 승강기 보유 건물용도 대분류
   제2종근린생활시설, 판매시설, 문화및집회시설, 공장, 업무시설, 숙박시설, 교육연구시설, 단독주택, 제1종근린생활시설, 창고시설, 방송통신시설, 공동주택, 노유자시설, 의료시설, 종교시설, 운수시설, 교정및군사시설, 운동시설, 수련시설, 발전시설, 관광휴게시설, 자동차관련시설, 위험물저장및처리시설, 묘지관련시설, 위락시설, 동식물관련시설, 자원순환관련시설, 야영장시설, 기타, 장례시설, 가설건축물}
   \item{정격속도_ms}{해당 승강기 정격속도}
   \item{대수}{해당 승강기 보유 건물의 승강기 대수}
}
}
\source{
\url{https://www.data.go.kr/data/15017366/fileData.do}
}
\usage{
elevator
}
\description{
한국승강기안전공단에서 제공하는 국내 승강기 보유 현황. 이용허락범위 제한 없음.
}
\examples{
DT::datatable(head(elevator,50))
}
\keyword{datasets}
