% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{소방청 화재통계}
\format{
A list with 18 data frame.
\describe{
   \item{연번}{해당 화재 연번}
   \item{사망}{해당 화재 사망}
   \item{부상}{해당 화재 부상}
   \item{인명피해.명.소계}{해당 화재 인명피해 소계}
   \item{재산피해소계}{해당 화재 재산피해소계}
   \item{시도}{해당 화재의 시 또는 도}
   \item{시군구}{해당 화재의 시 또는 군,구}
   \item{읍면동}{해당 화재의 읍 또는 면,동}
   \item{발화열원}{해당 화재의 발화열원
   작동기기, 담뱃불, 라이터불, 미상(발화원인), 불꽃, 불티, 마찰, 전도, 복사, 화학적 발화열, 폭발물, 폭죽, 기타(발화원인), 자연적 발화열}
   \item{발화열원소분류}{해당 화재의 발화열원소분류
   전기적 아크(단락), 담뱃불, 기기 전도,복사열, 미상, 모닥불, 연탄, 숯, 라이터불, 성냥불, 마찰열,마찰 스파크, 촛불, 기타(불꽃,불티), 화학반응열, 폭죽, 기타(작동기기), 굴뚝(연통) 아궁이, 불꽃, 스파크, 정전기, 화염 전도,복사열, 용접, 절단, 연마, 비화, 기타, 기타(마찰,전도,복사), 쓰레기, 논밭두렁, 기타(담뱃불,라이터불), 햇볕, 역화, 폭탄, 탄약, 향불, 낙뢰}
   \item{발화요인대분류}{해당 화재의 발화요인대분류
   전기적 요인, 부주의, 기계적 요인, 미상, 방화의심, 화학적 요인, 기타, 자연적인 요인, 교통사고, 방화, 가스누출(폭발)}
   \item{발화요인소분류}{해당 화재의 발화요인소분류
   접촉불량에 의한 단락, 담배꽁초, 음식물 조리중, 과열, 과부하, 미상, 불씨,불꽃,화원방치, 절연열화에 의한 단락, 방화의심, 미확인단락, 기타(화학적요인), 폭죽놀이, 자동제어 실패, 가연물 근접방치, 반단선, 기타(부주의), 기타, 압착,손상에 의한 단락, 기타(전기적요인), 과부하/과전류, 누전,지락, 용접, 절단 등}
   \item{최초착화물대분류}{해당 화재의 최초착화물대분류
   최초착화물 : 발화열원에 의해 불이 붙고 이 물질을 통해 제어하기 힘든 화세로 발전한 가연물
   전기,전자, 기타, 침구,직물류, 종이,목재,건초등, 식품, 자동차,철도차량,선박,항공기, 미상, 쓰레기류, 합성수지, 간판,차양막등, 가구, 위험물등, 가연성가스}
   \item{최초착화물소분류}{해당 화재의 최초착화물소분류
   최초착화물 : 발화열원에 의해 불이 붙고 이 물질을 통해 제어하기 힘든 화세로 발전한 가연물
   기타(종이,목재,건초등), 음식물, 벨트, 종이, 미상, 풀, 나뭇잎, 기타 쓰레기, 쓰레기, 건초, 합성고무(타이어), 톱밥, 기타(자동차,철도차량,선박,항공기), 목재, 합판, 전기, 전자기기 기판, 나무, 분진, 기타(간판,차양막등), 플라스틱, PVC, 비닐, 장판, 의류, 잔디, 부품, 타이어, 전기 등}
   \item{장소대분류}{해당 화재의 장소대분류
   주거, 임야, 자동차,철도차량, 판매,업무시설, 생활서비스, 기타, 산업시설, 기타서비스, 운수자동차시설, 집합시설, 교육시설, 의료,복지시설, 선박,항공기, 위험물,가스제조소, 문화재시설}
   \item{장소중분류}{해당 화재의 장소중분류
   단독주택, 들불, 자동차, 일반업무, 공동주택, 오락시설, 야외, 일상서비스, 공장시설, 판매시설, 음식점, 동식물시설, 산불, 위락시설, 기타건축물, 기타주택, 작업장, 도로, 창고시설, 자동차시설, 숙박시설, 종교시설, 학교, 의료시설, 노유자시설, 건설기계, 건강시설, 운동시설, 발전시설, 역사,터미널, 공공기관, 위생시설, 연구,학원 등}
   \item{장소소분류}{해당 화재의 장소소분류
   단독주택, 기타 들불, 오토바이, 일반빌딩, 다세대주택, 승용자동차, PC방(인터넷게임제공업), 기타야외, 기타 일상서비스, 쓰레기, 아파트, 금속기계 및 기구공업, 그 밖의 공업, 상점가, 커피숍,다방,카페(휴게), 노래연습장, 치킨,족발 등}
   \item{지번동}{해당 화재의 지번동}
}
}
\source{
\url{https://www.data.go.kr/data/3038724/fileData.do}{공공데이터포털:공공질서및안전 - 안전관리}
}
\usage{
fire
}
\description{
소방청에서 제공하는 장소별 원인별 화재통계.
Last Update: 2020-03-19;
}
\examples{
DT::datatable(head(fire,50))
}
\keyword{datasets}
