% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fireStation}
\alias{fireStation}
\title{소방청 전국 소방서 정보}
\format{
A data frame with 5 variables.
}
\usage{
fireStation
}
\description{
소방청에서 제공하는 전국 소방서 정보.
Last Update: 2022-04-18;
https://www.data.go.kr/data/15048242/fileData.do
}
\keyword{datasets}
