% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foodNutrients}
\alias{foodNutrients}
\title{식품의약품안전처 식품영양성분 데이터베이스}
\format{
A list with 146 data frame.
\describe{
   \item{no}{no}
   \item{sample_id}{sample_id}
   \item{식품코드}{식품코드}
   \item{db_군}{농축산물}
   \item{상용제품}{상용제품}
   \item{식품명}{식품명}
   \item{연도}{2019, 2020, 2021}
   \item{지역_제조사}{전국(대표)}
   \item{채취시기}{평균}
   \item{식품대분류}{농축산물}
   \item{식품상세분류}{곡류 및 그 제품, 감자 및 전분류, 당류, 두류, 견과류 및 종실류, 채소류, 버섯류, 과실류, 육류, 난류, 어패류 및 기타 수산물, 해조류, 우유 및 유제품류, 유지류, 차류, 음료류, 주류, 조미료류, 조리가공품류, 기타}
   \item{x1회제공량}{100}
   \item{내용량_단위}{g_그램}
   \item{에너지}{에너지}
   \item{수분_g}{수분_g}
   \item{단백질_g}{단백질_g}
   \item{지방_g}{지방_g}
   \item{탄수화물_g}{탄수화물_g}
   \item{총당류_g}{총당류_g}
   \item{자당_g}{자당_g}
   \item{포도당_g}{포도당_g}
   \item{과당_g}{과당_g}
   \item{유당_g}{유당_g}
   \item{맥아당_g}{맥아당_g}
   \item{갈락토오스_g}{갈락토오스_g}
   \item{총_식이섬유_g}{총 식이섬유_g}
   \item{수용성_식이섬유_g}{수용성 식이섬유_g}
   \item{불용성_식이섬유_g}{불용성 식이섬유_g}
   \item{칼슘}{칼슘}
   \item{철}{철}
   \item{마그네슘}{마그네슘}
   \item{인}{인}
   \item{칼륨}{칼륨}
   \item{나트륨}{나트륨}
   \item{아연}{아연}
   \item{구리}{구리}
   \item{망간}{망간}
   \item{셀레늄}{셀레늄}
   \item{몰리브덴}{몰리브덴}
   \item{요오드}{요오드}
   \item{레티놀}{레티놀}
   \item{베타카로틴}{베타카로틴}
   \item{비타민_d_d2_d3}{비타민 d_d2_d3}
   \item{비타민_d2}{비타민 d2}
   \item{비타민_d3}{비타민 d3}
   \item{비타민_e}{비타민 e}
   \item{알파_토코페롤}{알파 토코페롤}
   \item{베타_토코페롤}{베타 토코페롤}
   \item{감마_토코페롤}{감마 토코페롤}
   \item{델타_토코페롤}{델타 토코페롤}
   \item{알파_토코트리에놀}{알파 토코트리에놀}
   \item{베타_토코트리에놀}{베타 토코트리에놀}
   \item{감마_토코트리에놀}{감마 토코트리에놀}
   \item{델타_토코트리에놀}{델타 토코트리에놀}
   \item{비타민_k}{비타민 k}
   \item{비타민_k1}{비타민 k1}
   \item{비타민_k2}{비타민 k2}
   \item{비타민_b1}{비타민 b1}
   \item{비타민_b2}{비타민 b2}
   \item{나이아신}{나이아신}
   \item{나이아신_ne}{나이아신_ne}
   \item{니코틴산}{니코틴산}
   \item{니코틴아마이드}{니코틴아마이드}
   \item{판토텐산}{판토텐산}
   \item{비타민_b6}{비타민 b6}
   \item{피리독신}{피리독신}
   \item{비오틴}{비오틴}
   \item{엽산_dfe}{엽산 dfe}
   \item{엽산_천연_엽산}{엽산 천연 엽산}
   \item{엽산_합성_엽산}{엽산 합성 엽산}
   \item{비타민_b12}{비타민 b12}
   \item{비타민_c}{비타민 c}
   \item{총_아미노산}{총 아미노산}
   \item{필수_아미노산}{필수 아미노산}
   \item{이소류신}{이소류신}
   \item{류신}{류신}
   \item{라이신}{라이신}
   \item{메티오닌}{메티오닌}
   \item{페닐알라닌}{페닐알라닌}
   \item{트레오닌}{트레오닌}
   \item{트립토판}{트립토판}
   \item{발린}{발린}
   \item{히스티딘}{히스티딘}
   \item{아르기닌}{아르기닌}
   \item{비필수_아미노산}{비필수 아미노산}
   \item{티로신}{티로신}
   \item{시스테인}{시스테인}
   \item{알라닌}{알라닌}
   \item{아스파르트산}{아스파르트산}
   \item{글루탐산}{글루탐산}
   \item{글리신}{글리신}
   \item{프롤린}{프롤린}
   \item{세린}{세린}
   \item{타우린}{타우린}
   \item{콜레스테롤}{콜레스테롤}
   \item{총_지방산_g}{총 지방산_g}
   \item{총_필수_지방산_g}{총 필수 지방산_g}
   \item{총_포화_지방산_g}{총 포화 지방산_g}
   \item{부티르산_4_0}{부티르산 4_0}
   \item{카프로산_6_0}{카프로산 6_0}
   \item{카프릴산_8_0}{카프릴산 8_0}
   \item{카프르산_10_0}{카프르산 10_0}
   \item{라우르산_12_0}{라우르산 12_0}
   \item{트라이데칸산_13_0}{트라이데칸산 13_0}
   \item{미리스트산_14_0}{미리스트산 14_0}
   \item{펜타데칸산_15_0}{펜타데칸산 15_0}
   \item{팔미트산_16_0}{팔미트산 16_0}
   \item{헵타데칸산_17_0}{헵타데칸산 17_0}
   \item{스테아르산_18_0}{스테아르산 18_0}
   \item{아라키드산_20_0}{아라키드산 20_0}
   \item{헨에이코산산_21_0}{헨에이코산산 21_0}
   \item{베헨산_22_0}{베헨산 22_0}
   \item{트리코산산_23_0}{트리코산산 23_0}
   \item{리그노세르산_24_0}{리그노세르산 24_0}
   \item{총_단일_불포화지방산_g}{총 단일 불포화지방산_g}
   \item{미리스톨레산_14_1}{미리스톨레산 14_1}
   \item{팔미톨레산_16_1}{팔미톨레산 16_1}
   \item{헵타데센산_17_1}{헵타데센산 17_1}
   \item{올레산_18_1_n_9}{올레산 18_1_n_9}
   \item{박센산_18_1_n_7}{박센산 18_1_n_7}
   \item{가돌레산_20_1}{가돌레산 20_1}
   \item{에루크산_22_1}{에루크산 22_1}
   \item{네르본산_24_1}{네르본산 24_1}
   \item{총_다중_불포화지방산_g}{총 다중 불포화지방산_g}
   \item{리놀레산_18_2_n_6_c}{리놀레산 18_2_n_6_c}
   \item{알파_리놀렌산_18_3_n_3}{알파 리놀렌산_18_3_n_3}
   \item{감마_리놀렌산_18_3_n_6}{감마 리놀렌산_18_3_n_6}
   \item{에이코사디에노산_20_2_n_6}{에이코사디에노산 20_2_n_6}
   \item{에이코사트리에노산_20_3_n_3}{에이코사트리에노산 20_3_n_3}
   \item{에이코사트리에노산_20_3_n_6}{에이코사트리에노산 20_3_n_6}
   \item{아라키돈산_20_4_n_6}{아라키돈산 20_4_n_6}
   \item{에이코사펜타에노산_20_5_n_3}{에이코사펜타에노산 20_5_n_3}
   \item{도코사디에노산_22_2}{도코사디에노산 22_2}
   \item{도코사펜타에노산_22_5_n_3}{도코사펜타에노산 22_5_n_3}
   \item{도코사헥사에노산_22_6_n_3}{도코사헥사에노산 22_6_n_3}
   \item{오메가_3_지방산_g}{오메가 3 지방산_g}
   \item{오메가_6_지방산_g}{오메가 6 지방산_g}
   \item{트랜스_지방산_g}{트랜스 지방산_g}
   \item{트랜스_올레산_18_1_n_9_t}{트랜스 올레산_18_1_n_9_t}
   \item{트랜스_리놀레산_18_2t}{트랜스 리놀레산_18_2t}
   \item{트랜스_리놀렌산_18_3t}{트랜스 리놀렌산_18_3t}
   \item{식염상당량_g}{식염상당량_g}
   \item{회분_g}{회분_g}
   \item{폐기율_percent}{폐기율_percent}
   \item{성분표출처}{성분표출처}
   \item{발행기관}{발행기관}
}
}
\usage{
foodNutrients
}
\description{
식품의약품안전처에서 제공하는 농축수산물, 가공식품 및 음식 영양성분 데이터베이스. 이용허락범위: 저작자표시(CC BY). 
본 저작물은 '식품의약품안전처'에서 '2020년'작성하여 공공누리 제1유형으로 개방한 '식품영양성분 데이터베이스를'을 이용하였으며, 
해당 저작물은 '식품의약품안전처 식품영양성분 데이터베이스(https://www.foodsafetykorea.go.kr/fcdb/index.do)'에서 무료로 다운받으실 수 있습니다.
Last Update: 2020-10-28;
https://www.data.go.kr/data/15047698/fileData.do
}
\keyword{datasets}
