% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasStation}
\alias{gasStation}
\title{gasStation}
\format{
A data frame with 6 variables.
}
\usage{
gasStation
}
\description{
산업통상자원부에서 제공하는 전국 주유소 등록현황(전국 주유소 신규 등록 , 휴업, 폐업 현황 정보, 연도, 변동사유발생연월일, 판매업종류 등). 
Last Update: 2022-03-10;
https://www.data.go.kr/data/3076606/fileData.do
}
\keyword{datasets}
