% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasStation}
\alias{gasStation}
\title{산업통상자원부 전국 주유소 등록현황}
\format{
A data frame with 6 variables.
\describe{
   \item{연도}{연도}
   \item{변동사유발생연월일}{변동사유발생연월일}
   \item{판매업종류 }{휴업, 신규등록, 폐업, 등록취소}
   \item{구분}{구분}
   \item{업체명}{업체명}
   \item{소재지}{소재지}
 }
}
\usage{
gasStation
}
\description{
산업통상자원부에서 제공하는 전국 주유소 등록현황(전국 주유소 신규 등록 , 휴업, 폐업 현황 정보, 연도, 변동사유발생연월일, 판매업종류 등). 
Last Update: 2022-03-10;
https://www.data.go.kr/data/3076606/fileData.do
}
\keyword{datasets}
