% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalBusiness}
\alias{globalBusiness}
\title{대한무역투자진흥공사 해외진출기업 정보}
\format{
A data frame with 19 variables.
}
\usage{
globalBusiness
}
\description{
대한무역투자진흥공사에서 제공하는 전세계 124개국의 해외진출기업의 회사명, 주소, 홈페이지, 진출년도, 진출형태, 투자형태 등의 정보.
Last Update: 2021-09-27;
https://www.data.go.kr/data/15003297/fileData.do
}
\keyword{datasets}
