% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gyeonggiER}
\alias{gyeonggiER}
\title{경기도 응급의료기관 및 응급의료지원센터 현황}
\format{
A data frame with 14 variables.
\describe{
   \item{시군명}{해당 응급의료기관 및 응급의료지원센터의 시 또는 군}
   \item{병원명/센터명}{해당 응급의료기관 및 응급의료지원센터의 병원명 또는 센터명}
   \item{업무구분명}{해당 응급의료기관 및 응급의료지원센터의 업무구분명
   응급의료지원센터, 지역기관, 지역센터, 권역센터}
   \item{대표전화번호}{해당 응급의료기관 및 응급의료지원센터의 대표전화번호}
   \item{소재지도로명주소}{해당 응급의료기관 및 응급의료지원센터의 도로명주소}
   \item{소재지지번주소}{해당 응급의료기관 및 응급의료지원센터의 지번주소}
   \item{소재지우편번호}{해당 응급의료기관 및 응급의료지원센터의 우편번호}
   \item{위도}{해당 응급의료기관 및 응급의료지원센터의 위도}
   \item{경도}{해당 응급의료기관 및 응급의료지원센터의 경도}
   \item{응급의료지원센터여부}{해당 응급의료기관 및 응급의료지원센터의 응급의료지원센터여부
   Y or NA}
   \item{전문응급의료센터여부}{해당 응급의료기관 및 응급의료지원센터의 전문응급의료센터여부
   Y or NA}
   \item{전문응급센터전문분야}{해당 응급의료기관 및 응급의료지원센터의 전문응급센터전문분야
   소아전문응급실}
   \item{권역외상센터여부}{해당 응급의료기관 및 응급의료지원센터의 권역외상센터여부
   Y or NA}
   \item{지역외상센터여부}{해당 응급의료기관 및 응급의료지원센터의 지역외상센터여부
   Y or NA}
}
}
\source{
\url{https://www.data.go.kr/data/15011313/fileData.do}
}
\usage{
gyeonggiER
}
\description{
경기도 내 응급의료기관 및 응급의료지원센터 현황. 기관 및 센터의 소재지 주소와 대표 전화번호 등의 정보를 제공. 이용허락범위 제한 없음.
}
\examples{
library(leaflet)

gyeonggiER |> 
leaflet() |> 
  addTiles() |> 
  addMarkers(~경도, ~위도, popup = ~as.character(`병원명/센터명`), label = ~as.character(`병원명/센터명`)) |> 
  addMarkers(~병원경도, ~병원위도, popup = ~as.character(기관명), label = ~as.character(기관명), data = datatoys::seoulER)
DT::datatable(head(gyeonggiER,50))
}
\keyword{datasets}
