% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitalInfo}
\alias{hospitalInfo}
\title{건강보험심사평가원 병의원 기본정보}
\format{
A data frame with 29 variables.
\describe{
   \item{암호화요양기호}{암호화요양기호}
   \item{요양기관명}{해당 병의원 기관명}
   \item{종별코드}{해당 병의원 종별코드}
   \item{종별코드명}{해당 병의원 종별코드명
   상급종합, 종합병원, 병원, 요양병원, 정신병원, 의원}
   \item{시도코드}{해당 병의원 시도코드}
   \item{시도코드명}{해당 병의원 시도코드명
   인천, 강원, 서울, 대구, 경남, 경기, 충남, 부산, 전북, 광주, 대전, 충북, 울산, 전남, 경북, 세종시, 제주}
   \item{시군구코드}{해당 병의원 시 또는 군,구 코드}
   \item{시군구코드명}{해당 병의원 시 또는 군,구 코드명}
   \item{읍면동}{해당 병의원 읍 또는 면, 동}
   \item{우편번호}{해당 병의원 우편번호}
   \item{주소}{해당 병의원 주소}
   \item{전화번호}{해당 병의원 전화번호}
   \item{병원URL}{해당 병의원 병원URL}
   \item{개설일자}{해당 병의원 개설일자}
   \item{총의사수}{해당 병의원 총 의사수}
   \item{의과일반의.인원수}{해당 병의원 의과일반의 인원수}
   \item{의과인턴.인원수}{해당 병의원 의과인턴 인원수}
   \item{의과레지던트.인원수}{해당 병의원 의과레지던트 인원수}
   \item{의과전문의.인원수}{해당 병의원 의과전문의 인원수}
   \item{치과일반의.인원수}{해당 병의원 치과일반의 인원수}
   \item{치과인턴.인원수}{해당 병의원 치과인턴 인원수}
   \item{치과레지던트.인원수}{해당 병의원 치과레지던트 인원수}
   \item{치과전문의.인원수}{해당 병의원 치과전문의 인원수}
   \item{한방일반의.인원수}{해당 병의원 한방일반의 인원수}
   \item{한방인턴.인원수}{해당 병의원 한방인턴 인원수}
   \item{한방레지던트.인원수}{해당 병의원 한방레지던트 인원수}
   \item{한방전문의.인원수}{해당 병의원 한방전문의 인원수}
   \item{x좌표}{해당 병의원 x좌표}
   \item{y좌표}{해당 병의원 y좌표}
 }
}
\source{
\url{https://www.data.go.kr/data/15051059/fileData.do}{공공데이터포털:보건 - 건강보험}
}
\usage{
hospitalInfo
}
\description{
건강보험심사평가원에서 제공하는 병의원 기본정보 데이터셋.
Last Update: 2022-04-14;
}
\examples{
DT::datatable(head(hospitalInfo,50))
}
\keyword{datasets}
