% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housingPrice}
\alias{housingPrice}
\title{국토교통부 2021 공동주택 공시가격 정보}
\format{
A data frame with 20 variables.
\describe{
   \item{기준연도}{해당 정보의 기준연도}
   \item{기준월}{해당 정보의 기준월}
   \item{법정동코드}{해당 공동주택의 법정동코드
   법정동 : 법률로 지정한 행정구역 단위
  법정동코드 : 10자리로 구성, 광역시(2자리) + 시군구(3자리) + 읍면동(3자리) + 리(2자리)
  (예) 서울특별시(11) + 용산구(170) + 한남동(131) + 리(00) = 1117013100}
   \item{도로명주소}{해당 정보의 도로명주소
   도로명주소 : 도로에 이름을 붙이고 주택, 건물에는 도로를 따라 순차적으로 번호를 붙여 도로명과 건물번호에 의해 표기하는 주소}
   \item{시도}{해당 공동주택의 시 또는 도}
   \item{시군구}{해당 공동주택의 시 또는 군, 구}
   \item{읍면}{해당 공동주택의 읍 또는 면}
   \item{동리}{해당 공동주택의 동 또는 리}
   \item{특수지코드}{해당 공동주택의 특수지코드
   특수지 : 토지용도가 특수하고 거래사례가 희소하여 시장가치 측정이 어려운 토지}
   \item{본번}{해당 정보의 도로명주소의 본번
   (예) 서울특별시 노원구 덕릉로10길 30-20, 801동 102호 : 본번 30}
   \item{부번}{해당 정보의 도로명주소의 부번
   (예) 서울특별시 노원구 덕릉로10길 30-20, 801동 102호 : 부번 20}
   \item{특수지명}{해당 정보의 도로명주소의 특수지명
   특수지 : 토지용도가 특수하고 거래사례가 희소하여 시장가치 측정이 어려운 토지}
   \item{단지명}{해당 정보의 도로명주소의 단지명}
   \item{동명}{해당 정보의 도로명주소의 동명}
   \item{호명}{해당 정보의 도로명주소의 호명}
   \item{전용면적}{해당 정보의 도로명주소의 전용면적}
   \item{공시가격}{해당 정보의 도로명주소의 공시가격}
   \item{단지코드}{해당 정보의 도로명주소의 단지코드}
   \item{동코드}{해당 정보의 도로명주소의 동코드}
   \item{호코드}{해당 정보의 도로명주소의 호코드}
 }
}
\source{
\url{https://www.data.go.kr/data/3073746/fileData.do}
}
\usage{
housingPrice
}
\description{
국토교통부에서 제공하는 2021년도 공동주택 공시가격정보입니다.
}
\examples{
DT::datatable(head(housingPrice, 50))
}
\keyword{datasets}
