% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housingPrice}
\alias{housingPrice}
\title{국토교통부 2021 공동주택 공시가격 정보}
\format{

}
\source{
\url{https://www.data.go.kr/data/3073746/fileData.do}
}
\usage{
housingPrice
}
\description{
국토교통부에서 제공하는 2021년도 공동주택 공시가격정보입니다.
}
\examples{
DT::datatable(head(housingPrice, 50))
}
\keyword{datasets}
