% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{karaoke}
\alias{karaoke}
\title{행정안전부 단란주점 영업 정보}
\format{
A list with 48 data frame.
}
\usage{
karaoke
}
\description{
행정안전부에서 제공하는 단란주점 영업 정보. 주로 주류와 함께 음식류를 조리 및 판매하는 곳으로 손님이 노래를 부르는 행위가 허용되는 업소정보 데이터로 인허가일자, 영업상태, 사업장명, 소재지주소 등의 정보등을 제공. * 좌표계 : 중부원점TM(EPSG:2097).
Last Update: 2021-11-19;
https://www.data.go.kr/data/15045017/fileData.do
}
\keyword{datasets}
