% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legalDong}
\alias{legalDong}
\title{legalDong}
\format{
A data frame with 9 variables.
}
\usage{
legalDong
}
\description{
국토교통부에서 제공하는 법정동 정보. 우리나라의 법정 구역으로 법률로 지정된 일정한 명칭과 영역을 지닌 구역으로 
토지행정시스템에서 사용하는 법정동 데이터. 현존하는 법정동코드는 삭제일자가 없으며, 과거법정동코드는
행정표준코드관리시스템(https://www.code.go.kr)에서 제공하는 법정동코드 자료를 바탕으로 시스템 사용 용도에 
맞춰 가공한 데이터로 정보가 상이할 수 있음.
Last Update: 2021-12-17;
https://www.data.go.kr/data/15063424/fileData.do
}
\keyword{datasets}
