% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicalCheckup}
\alias{medicalCheckup}
\title{국민건강보험 일반건강검진결과}
\format{
A data frame with 31 variables.
\describe{
 \item{기준년도}{해당 정보의 기준년도}
 \item{가입자.일련번호}{해당가입자에 부여한 일련번호}
 \item{시도코드}{해당 수진자(진료를 받은 사람) 거주지의 시, 도 코드
 코드 11 : 서울특별시, 코드 26 : 부산광역시, 코드 27 : 대구광역시, 코드 28: 인천광역시, 코드 29 : 광주광역시, 코드 30 : 대전광역시, 코드 31 : 울산광역시, 코드 36 : 세종특별자치시, 코드 41 : 경기도, 코드 42 : 강원도, 코드 43 : 충청북도, 코드 44 : 충청남도, 코드 45: 전라북도, 코드 46: 전라남도, 코드 47 : 경상북도, 코드 48 : 경상남도, 코드 49 : 제주특별자치도}
 \item{성별코드}{해당 정보 대상자의 성별, 1 : 남자, 2 : 여자}
 \item{연령대.코드.5세단위.}{기준년도에 수진자(진료를 받은 사람)의 나이를 5세 단위로 그룹화(범주화)하여 구분한 코드
 그룹 1: 0세~4세, 그룹 2: 5세 ~ 9세, 그룹 3: 10세 ~ 14세, 그룹 4 : 15세 ~ 19세, 그룹 5 : 20세 ~ 24세, 그룹 6 : 25세 ~ 29세, 그룹 7 : 30세 ~ 34세, 그룹 8: 35세 ~ 39세, 그룹 9 : 40세 ~ 44세, 그룹 10 : 45세 ~ 49세, 그룹 11: 50세 ~ 54세, 그룹 12 : 55세 ~ 59세, 그룹 13 : 60세 ~ 64세, 그룹 14 : 65세 ~ 69세, 그룹 15 : 70세 ~ 74세, 그룹 16 : 75세 ~ 79세, 그룹  17 :  80세 ~ 84세, 그룹 18 : 85세 이상}
 \item{신장.5Cm단위.}{검진자의 키를 5cm로 맞춘 단위
 (예) 100~104cm = 100cm}
 \item{체중.5Kg.단위.}{검진자의 몸무게를 5kg로 맞춘 단위
 (예) 70~74kg = 70kg}
 \item{허리둘레}{검진자의 허리둘레(기준년도 2002년~20007년까지의 데이터는 측정되지않아 결측 처리 돼 제공됨)}
 \item{시력.좌.}{수검자(진료를 받은 사람)의 좌측 눈 시력
 0.1~2,5의 값으로 표기하며 0.1 이하의 시력은 0.1로 표기하고 실명은 9.9로 표기}
 \item{시력.우.}{수검자(진료를 받은 사람)의 우측 눈 시력
 0.1~2,5의 값으로 표기하며 0.1 이하의 시력은 0.1로 표기하고 실명은 9.9로 표기}
 \item{청력.좌.}{수검자(진료를 받은 사람)의 좌측 귀의 청력
  1 : 정상, 2 : 비정상}
 \item{청력.우.}{수검자(진료를 받은 사람)의 우측 귀의 청력
  1 : 정상, 2 : 비정상}
 \item{수축기.혈압}{검진자의 최고 혈압으로 심장이 수축해서 강한 힘으로 혈액을 동맥에 보낼 때의 혈관 내압}
 \item{이완기.혈압}{검진자의 최저 혈압으로 심장의 완기시의 혈압}
 \item{식전혈당.공복혈당.}{검진자 식전 전 혈당(혈액 100ml당 함유 되어 있는 포도당의 농도) 수치}
 \item{총.콜레스테롤}{혈청 중의 에스텔형, 비에스테형(유리)콜레스테롤의 합
 정상치는 150~250mg/dL 약 1/3이 비에스텔형(유리)콜레스테롤이며 나머지가 콜레스테롤에스테르}
 \item{트리글리세라이드}{단순지질 혹은 중성지질을 뜻함
 글리세롤에 3분자 지방산이 에스테르 합한 것으로서 자연계에서 찾아낼 수 있는 지방유도체 가운데 가장 분포가 넓음
 정상치는 30~135mg/dL(0.34~1.52-mmol/ℓ)(기준년도 2002년~20007년까지의 데이터는 측정되지않아 결측 처리 돼 제공됨)}
 \item{HDL.콜레스테롤}{HDL(고밀도 리포단백질)에 포함되는 콜레스테롤
 작은 입자의 콜레스테롤로 세포에 이끌려간 콜레스테롤을 간으로 돌려주고 혈관 벽에 쌓인 나쁜 콜레스테롤을 없애는 역할을 하는 성분
 정상치는 30~65mg/dL(기준년도 2002년~20007년까지의 데이터는 측정되지않아 결측 처리 돼 제공됨)}
 \item{LDL.콜레스테롤}{LDL(저밀도 리포단백질)에 함유된 콜레스테롤
 입자가 매우 큰 콜레스테롤로 양이 과도하게 증가할 경우,혈관벽에 쌓여서 동맥경화나 각종 질병을 야기 하는 성분
 170mg/dL 이상일 경우 일반적으로 고LDL혈증으로 봄(기준년도 2002년~20007년까지의 데이터는 측정되지않아 결측 처리 돼 제공됨)}
 \item{혈색소}{혈액이나 혈구 속에 존재하는 색소단백으로 글로빈(globin)과 헴(heme)으로 구성되며 혈중의 산소운반체로서의 역할 수행}
 \item{요단백}{변에 단백질이 섞여 나오는 것
 1(-), 2(±), 3(+1), 4(+2), 5(+3), 6(+4)로 표기됨}
 \item{혈청크레아티닌}{크레아티닌은 크레아틴의 탈수물로 내인성 단백대사의 종말산물로서 신장에서 배설되고 그 증감은 음식물에 관계없이 근육의 발육과 운동에 관계함, 혈청크레아티닌 농도는 신기능장애에 의해 증량함
 정상치 0.8~1.7mg/dL(기준년도 2002년~20007년까지의 데이터는 측정되지않아 결측 처리 돼 제공됨)}
 \item{X.혈청지오티.AST}{간 기능을 나타내는 혈액검사상의 수치
 간세포 이외에 심장, 신장, 뇌, 근육 등에도 존재하는 효소로 이러한 세포들이 손상을 받는 경우 농도가 증가함
 정상치 0~40IU/L}
 \item{X.혈청V지오티.ALT}{간 기능을 나타내는 혈액검사상의 수치
 ALT는 주로 간세포 안에 존재하는 효소로, 간세포가 손상을 받는 경우 농도가 증가함
 정상치 0~40IU/L}
 \item{감마.지티피}{ 기능을 나타내는 혈액검사상의 수치
 간 내의 쓸개관(담관)에 존재하는 효소로 글루타민산을 외부에 펩티드나 아미노산 등으로 옮기는 작용을 함
 쓸개즙(담즙) 배설 장애, 간세포 장애 발생 시 혈중에 증가하게 됨
 정상치 남성 11~63IU/L, 여성 8~35IU/L}
 \item{흡연상태}{해당 수검자(진료를 받은 사람)의 흡연 상태 여부
 1 : 피우지 않는다, 2 : 이전에 피웠으나 끊었다, 3 : 현재도 피우고 있다}
 \item{음주여부}{해당 수검자(진료를 받은 사람)의 음주 상태 여부
 0 : 마시지않는다, 1 : 마신다.}
 \item{구강검진.수검여부}{해당 검진자가 구강검진을 선택하여 검진하였는지 여부에 대한 항목
 0 : 미수검, 1 : 수검}
 \item{치아우식증유무}{해당 수검자(진료를 받은 사람)의 치우우식증 유무에 대한 항목
 0 : 없음,1 : 있음}
 \item{치석}{해당 수검자의 치석 여부
 치석 : 치근 부위의 잇몸이 내려가 노출된 치아표면이 마모된 것으로 잇몸이 내려앉아 시멘트질 안의 상아질이 노출된 상태
 0(없음), 1(있음)}
 \item{데이터.공개일자}{데이터 작성 기준일자}
}
}
\source{
\url{https://www.data.go.kr/data/15007122/fileData.do}{공공데이터포털:보건 - 보건의료}
}
\usage{
medicalCheckup
}
\description{
국민건강보험의 직장가입자와 40세 이상의 피부양자, 세대주인 지역가입자와 40세 이상의 지역가입자의 일반건강검진 결과와 이들 일반건강검진 
대상자 중에 만40세와 만66세에 도달한 이들이 받게 되는 생애전환기건강진단 수검이력이 있는 각 연도별 수진자 100만 명에 대한 
기본정보(성, 연령대, 시도코드 등)와 검진내역(신장, 체중, 총콜레스테롤, 혈색소 등)으로 구성된 개방데이터
}
\examples{
DT::datatable(head(medicalCheckup, 50))
}
\keyword{datasets}
