% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicine}
\alias{medicine}
\title{건강보험심사평가원 의약품 주성분 정보}
\format{
A data frame with 8 variables.
}
\usage{
medicine
}
\description{
건강보험심사평가원에서 제공하는 의약품 주성분(일반명) 현황에 대한 데이터셋. 
의약품주성분 코드, 성분명, 제형구분, 투여, 함량, 단위 등 정보를 제공.
Last Update: 2021-10-25;
https://www.data.go.kr/data/15067461/fileData.do
}
\keyword{datasets}
