% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nationalPension}
\alias{nationalPension}
\title{국민연금공단 국민연금사업장 정보}
\format{
A data frame with 22 variables.
\describe{
   \item{자료생성년월}{자료생성년월}
   \item{사업장명}{사업장명}
   \item{사업자등록번호}{사업자등록번호}
   \item{사업장가입상태코드.1.등록.2.탈퇴}{사업장가입상태코드.1.등록.2.탈퇴}
   \item{우편번호}{우편번호}
   \item{사업장지번상세주소}{사업장지번상세주소}
   \item{사업장도로명상세주소}{사업장도로명상세주소}
   \item{고객법정동주소코드}{고객법정동주소코드}
   \item{고객행정동주소코드}{고객행정동주소코드}
   \item{법정동주소광역시도코드}{법정동주소광역시도코드}
   \item{법정동주소광역시시군구코드}{법정동주소광역시시군구코드}
   \item{법정동주소광역시시군구읍면동코드}{법정동주소광역시시군구읍면동코드}
   \item{사업장형태구분코드.1.법인.2.개인}{사업장형태구분코드.1.법인.2.개인}
   \item{사업장업종코드}{사업장업종코드}
   \item{사업장업종코드명}{사업장업종코드명}
   \item{적용일자}{적용일자}
   \item{재등록일자}{재등록일자}
   \item{탈퇴일자}{탈퇴일자}
   \item{가입자수}{가입자수}
   \item{당월고지금액}{당월고지금액}
   \item{신규취득자수}{신규취득자수}
   \item{상실가입자수}{상실가입자수}
 }
}
\usage{
nationalPension
}
\description{
국민연금공단에서 제공하는 법정동단위 지역별, 국민연금 가입 사업장 정보(단, 개인사업장 및 2인 이하 법인 사업장 정보 미제공).
Last Update: 2022-05-18;
https://www.data.go.kr/data/15083277/fileData.do
}
\keyword{datasets}
