% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nationalPension}
\alias{nationalPension}
\title{국민연금공단 국민연금사업장 정보}
\format{
A data frame with 22 variables.
\describe{
   \item{자료생성년월}{자료생성년월
   월(영어).년(숫자)으로 기입}
   \item{사업장명}{국민연금 가입 사업장명}
   \item{사업자등록번호}{국민연금 가입 사업장 사업자등록번호}
   \item{사업장가입상태코드.1.등록.2.탈퇴}{해당 사업장 국민연금 가입 상태코드
   1.등록.2.탈퇴}
   \item{우편번호}{해당 사업장 우편번호}
   \item{사업장지번상세주소}{해당 사업장 지번 상세주소}
   \item{사업장도로명상세주소}{해당 사업장 도로명 상세주소}
   \item{고객법정동주소코드}{해당 고객 법정동주소코드
   법정동 : 법률로 지정한 행정구역 단위
   법정동코드 : 10자리로 구성, 광역시(2자리) + 시군구(3자리) + 읍면동(3자리) + 리(2자리)
   (예) 서울특별시(11) + 용산구(170) + 한남동(131) + 리(00) = 1117013100}
   \item{고객행정동주소코드}{해당 고객 행정동주소코드
   행정동주소 : 행정상 효율을 위해 인구 증감 및 이동 등에 맞춰 나눈 구역}
   \item{법정동주소광역시도코드}{해당 법정동주소 광역시도코드
   광역시 : 2자리}
   \item{법정동주소광역시시군구코드}{해당 법정동주소 광역시시군구코드
   시군구 : 3자리}
   \item{법정동주소광역시시군구읍면동코드}{법정동주소 광역시 시군구 읍면동 코드
   읍면동 :  3자리}
   \item{사업장형태구분코드.1.법인.2.개인}{국민연금 가입 사업장 구분 코드
   1 : 법인 2:개인}
   \item{사업장업종코드}{해당 사업장 업종코드
   업종코드 : 국세청이 세원관리 등을 위해 일정한 규칙에 따라 업종마다 부여한 번호}
   \item{사업장업종코드명}{해당 사업장 업종코드명}
   \item{적용일자}{해당 사업장 적용일자}
   \item{재등록일자}{국민연금 재등록한 사업장의 국민연금 재등록일자}
   \item{탈퇴일자}{국민연금 탈퇴한 사업장의 국민연금 탈퇴일자}
   \item{가입자수}{사업장내 국민연금 가입자수}
   \item{당월고지금액}{해당 사업장 국민연금 당월고지금액}
   \item{신규취득자수}{해당 사업장 국민연금 신규 취득자수}
   \item{상실가입자수}{해당 사업장 국민연금 상실 가입자수}
 }
}
\source{
\url{https://www.data.go.kr/data/15083277/fileData.do}
}
\usage{
nationalPension
}
\description{
국민연금공단에서 제공하는 법정동단위 지역별, 국민연금 가입 사업장 정보(단, 개인사업장 및 2인 이하 법인 사업장 정보 미제공).
}
\examples{
DT::datatable(head(nationalPension, 50))
}
\keyword{datasets}
