% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{necessariesPrice}
\alias{necessariesPrice}
\title{necessariesPrice}
\format{
A data frame with 7 variables.
}
\usage{
necessariesPrice
}
\description{
한국소비자원 참가격정보서비스에 접수된 상품별 생필품 가격 데이터 현황(상품명, 판매업소, 판매가격 등).
Last Update: 2022-05-23;
https://www.data.go.kr/data/15083256/fileData.do
}
\keyword{datasets}
