% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{odaNews}
\alias{odaNews}
\title{한국국제협력단 국가별 개발협력동향정보}
\format{
A data frame with 16 variables.
}
\usage{
odaNews
}
\description{
한국국제협력단에서 제공하는 국가별 개발 협력 동향정보 데이터. 지역, 국가, 관할사무소, 구분(공여국/수원국/국제기구/KOICA), 
KOICA 12대분야 및 기타, 동향기사 제목 및 본문내용, 출처, 링크 정보를 제공. 
Last Update: 2022-06-29;
https://www.data.go.kr/data/15052910/fileData.do
}
\keyword{datasets}
