% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{국립환경과학원 축산오염원조사정보}
\format{
A data frame with 18 variables: \code{법정동코드}, \code{시도},
  \code{시군구}, \code{읍면동}, \code{리}, \code{축종}, \code{법적규제}, \code{사육두수.두.},
  \code{농가수.호.}, \code{가축분뇨발생량.m3.일.}, \code{자가처리_정화방류.m3.일.}, \code{자가처리_퇴비화.m3.일.}, 
  \code{자가처리_액비화.m3.일.}, \code{미처리.m3.일.}, \code{위탁처리_공공처리시설.m3.일.}, \code{위탁처리_공동처리시설.m3.일.}, 
  \code{위탁처리_재활용신고자.m3.일.}, and \code{위탁처리_가축분뇨처리업.m3.일.}.
}
\usage{
pollution
}
\description{
국립환경과학원에서 제공하는 축산계 오염원에 대한 조사정보
https://www.data.go.kr/data/3045217/fileData.do
}
\keyword{datasets}
