% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scholarship}
\alias{scholarship}
\title{한국장학재단 2020년도 장학금 수혜현황}
\format{
A data frame with 16 variables.
\describe{
  \item{대학명}{장학금 지급 대학명}
  \item{학제별}{장학금 지급 대학의 학제구분
  전문대학, 대학}
  \item{설립별}{장학금 지급 대학의 설립 구분
  사립, 국공립}
  \item{지역별}{장학금 지급 대학의 지역 구분}
  \item{교외장학금.소계.원.}{교외장학금(국가, 지방자치단체, 사설및기타) 합계}
  \item{교외장학금.국가}{국가 재원 교외장학금}
  \item{교외장학금.지방자치단체}{지방자치단체 재원 교외장학금}
  \item{교외장학금.사설및기타}{사설 및 기타 재원 교외 장학금}
  \item{교내장학금.소계.원.}{교내장학금(성적우수, 저소득층, 근로, 재난, 교직원, 기타) 합계}
  \item{교내장학금.성적우수장학금}{학업성적이 우수한 자에게 지급한 교내장학금}
  \item{교내장학금.저소득층장학금}{경제적 사정이 어려운 자에게 지급한 교내 장학금}
  \item{교내장학금.근로장학금}{교내외 특정 장소에서 일정시간 근로한 자에게 지급한 교내장학금}
  \item{교내장학금.재난장학금}{재난을 사유로 지급한 교내장학금}
  \item{교내장학금.교직원장학금"}{대학에서 교직원 또는 그 가족에게 지급한 교내장학금}
  \item{교내장학금.기타}{5개 사유(성적우수, 저소득층, 근로, 재난, 교직원)에 해당하지 않는 교내 장학금}
  \item{총계.원.}{교외 장학금 및 교내장학금 합계}
  }
}
\source{
\url{https://www.data.go.kr/data/15038576/fileData.do}
}
\usage{
scholarship
}
\description{
한국장학재단에서 제공하는 전년도 장학금 수혜현황(대학정보공시 항목) 조사를 통해, 대학별 재원구분에 따른 교외장학금 수혜 금액 및 
성격유형에 따른 교내장학금 수혜금액 제공. 2020 회계연도부터 재난을 사유로 지급한 교내장학금 추가 조사.
}
\examples{
DT::datatable(head(scholarship, 50))
}
\keyword{datasets}
