% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scholarship}
\alias{scholarship}
\title{scholarship}
\format{
A data frame with 16 variables.
}
\usage{
scholarship
}
\description{
한국장학재단에서 제공하는 전년도 장학금 수혜현황(대학정보공시 항목) 조사를 통해, 대학별 재원구분에 따른 교외장학금 수혜 금액 및 
성격유형에 따른 교내장학금 수혜금액 제공. 2020 회계연도부터 재난을 사유로 지급한 교내장학금 추가 조사.
https://www.data.go.kr/data/15038576/fileData.do
}
\keyword{datasets}
